==> report/01-intro.Rmd <==
---
title: "导论"
author: "Jin"
date: "2019-04-15"
css: ./style/markdown.css
autoEqnLabels: true
eqnPrefixTemplate: ($$i$$)
linkReferences: true
bibliography: Bibfile.bib
notice: '@*'
csl: ./style/chinese-gb7714-2005-numeric.csl
link-citations: true
---


```{r setup, echo=F}

################# 第 1 章 R 程序代码  ####################


knitr::opts_knit$set(root.dir = getwd())
knitr::opts_chunk$set(echo = FALSE, results = 'hide')
knitr::opts_chunk$set(warning = FALSE, message=FALSE)
```

```{r prepare}
rm(list=ls())
options(digits=4)
options(scipen=100)
graphics.off()
Sys.setlocale("LC_ALL", "Chinese")
```

# 导论
## 选题背景

统计学中有很多模型，其中应用最为广泛的就是回归模型。通常我们使用回归模型去分析某
些变量对另一变量的影响。回归分析种类很多，可以根据具体研究目的及数据特点选择合适
的模型。经典线性回归模型可以用来研究某些变量对另一变量的线性影响，是最简单的回归
模型，但其有较为严格的假设条件。在实际应用中，数据不一定满足全部的假设条件。当数
据违背经典线性回归模型的假设时，继续使用经典线性回归模型是不合适的。这时可以引入
新的模型来解决违背经典线性回归模型假设的情况。广义线性模型的引入解决了离散因变量
的问题，可加模型的引入解决了自变量对因变量的非线性作用。在实际研究中，有时还会遇
到这样一些数据，其结构表现为在不同水平（或组）之间的观测值是相互独立的，而在同一
水平（或组）之间的观测值是相关的，即所研究的数据具有一定的异质性。针对这种情况，
可以通过在广义可加模型中引入随机效应来反映这些异质性，这就得到广义可加混合模型
（GAMM, Generalized Additive Mixed Model）。

在建立模型一般形式后，我们需要通过一定的方法对模型的参数进行估计，探究各自变量对
因变量的影响程度大小。在GAMM模型中，需要估计的参数包括固定效应的线性和非线性部分
及随机效应的方差成分。在统计学中，参数估计方法可以分为频率学派估计方法和贝叶斯学
派估计方法。针对参数估计，不同学派有不同的观点，针对贝叶斯学派的批评主要集中于参
数看成随机变量是否合适；先验分布是否存在以及如何选取。针对经典学派的批评主要集中
于一些统计推断问题的提法不妥，比如置信区间和假设检验；判断统计方法好坏的标准不妥
等[@张尧庭1991]。在解决实际问题时，各派有各派的优缺点，这就需要我们找到一个更加
符合实际情况的方法去解决相应的实际问题。针对GAMM，频率推断是基于边际似然原理，为
计算条件似然值，需要求似然函数的积分，而这些积分又常常是高维积分，高维积分在计算
上比较困难。这时一般解决方法有两种：一种是求近似积分，另一种是使用随机模拟的方法。
但这两种方法计算强度比较大。随着计算机计算能力的提升和相关程序的推出，广义可加混
合模型的参数估计问题逐步得到了解决。惩罚拟似然法（penelised quasi-likelihood，
PQL）就是经常采用的一种求近似积分的方法。但在使用过程中，基于PQL的估计逐渐暴露出
一些不足，Breslow(1995)指出PQL估计是有偏估计，其估计精度值得怀疑[@Breslow1993]。
另一种方法就是基于贝叶斯框架的马尔可夫链蒙特卡罗（Markov Chain Monte Carlo，MCMC）
随机模拟方法。后验分布是贝叶斯推断的本质，但是，通过贝叶斯理论找到后验分布并不是
那么容易,特别是在变量参数较多的时候，计算后验密度就比较困难。MCMC方法的出现，使
贝叶斯统计推断得到了进一步的发展。正如Bolstad（2017）指出的：MCMC方法用于解决难
以直接抽样的分布的随机抽样模拟问题，可以用在有很多参数的复杂模型中（Bolstad，
2017）[@BOLSTAD2017]。在MCMC中，比较常用的抽样方法有MH(Metropolis–Hastings)抽样、
Gibbs抽样，其他抽样方法都是基于这两种方法发展而来，比如切片抽样。MH抽样解决了
MCMC中接受率过低的问题，但在状态转移过程中，数据特征非常多时，较低的接受率会导致
算法效率较低，而且在高维的情况下，有时很难求出特征的联合分布，但条件概率分布相对
容易获得。因此，在MH算法的基础上产生了Gibbs抽样，其基本思想是对高维参数进行后验
推断时，通过参数向量某一个分量的条件分布来构造马尔可夫链，使得其平稳分布恰为目标
分布。在某些条件分布难以抽取而使Gibbs抽样无法进行的时候，一种替代方法就是使用切
片抽样。切片抽样对于单变量分布很容易实施，其基本思想是通过引入辅助变量，在两变量
的联合分布中抽取样本，之后忽略辅助变量。当然也可以通过依次更新每个变量从多变量的
分布中抽取样本，适用于各种类型的分布。通常来讲，切片抽样比Gibbs抽样实施起来更加
简单，比MH抽样更加有效（Neal，2003）[@Neal2003a]。考虑到切片抽样这样优点，一种基
于切片抽样的MCMC方法：sliceMCMC可以用于广义可加混合模型的参数估计。

随着广义可加混合模型参数估计方法的发展及统计软件对这些方法的实现，快速估计广义可
加混合模型已成为可能。广义可加混合模型可以应用到多个实际领域。但与理论研究相比，
目前广义可加混合模型的实际应用相对较少，国内广义可加混合模型的应用更少。因此，本
项目考虑GAMM模型在环境问题上的一个重要应用：空气质量问题。近几年来，空气质量问题
引起了社会各界的广泛关注，城市空气污染已经成为了一个严重的社会问题。根据世界卫生
组织在瑞士日内瓦发布的全球空气质量最新报告（2018）显示，空气污染在包括中国在内的
全球很多地方仍然处于危险水平。报告中最新数据显示，全球每10人中有9人呼吸的空气中
含有高浓度的污染物，并且揭示了2016年室内和室外空气污染导致全球大约700万人死亡。
报告估计，中国每年有将近200万人因环境和室内空气中的颗粒物造成的污染而死亡，而在
这些死亡中，世界卫生组织估计2016年有超过100万人死于环境空气污染 ^[资料来源于世界
卫生组织网站：http://www.wpro.who.int/china/mediacentre/]。中国国内生产总值已经
位居全球第二，但是前几年中国粗放型的经济发展模式带来的环境污染问题却日益严重，空
气质量也在逐年恶化。恶劣的空气环境不只严重影响了人们的健康和正常生活，也给社会各
行各业带来了巨大的经济损失。在一味地追求经济高速增长的同时，空气质量状况的日益下
降警醒我们比须及时采取相关措施提升城市空气质量。在2018年召开的全国生态环境保护大
会上，习近平总书记就提出要加大力度推进生态文明建设、解决生态环境问题，坚决打好污
染防治攻坚战，并要求坚决打赢蓝天保卫战是重中之重，要以空气质量明显改善为刚性要求，
基本消除重污染天气，还老百姓蓝天白云。要做到这些，就需要我们找到空气质量的影响因
素，进而根据影响因素及影响程度的大小进行相应的处理。因此，本项目利用广义可加混合
模型对城市空气质量影响因素从几种不同角度进行研究。

## 选题意义

GAMM模型可以很好的处理各种类型的因变量，但较为复杂的参数估计问题是限制其广泛使用
的主要原因。随着贝叶斯方法越来越成熟，计算机MCMC算法的应用，贝叶斯方法应用范围越
来越广泛，同时也解决了包括GAMM模型在内的很多模型的参数估计问题。本项目对GAMM模型
基于切片抽样MCMC估计和PQL估计的参数估计方法进行讨论，并根据空气质量及相关数据建
立GAMM模型，其意义主要体现在以下两个方面。

### 理论意义

广义可加混合模型可以很好的处理各种类型的自变量，既考虑了自变量对因变量的线性影响，
又考虑了非线性影响，同时还考虑了随机效应。在数据分析时，其假定条件相对较少，模型
设定可能更为精确。就估计方法来看，同时对用基于切片抽样的MCMC方法及传统的PQL方法
进行讨论，为实际使用GAMM进行数据分析提供了基础。由于GAMM模型相对较为复杂，实际使
用中也较为灵活，将GAMM应用于空气质量研究，展示该模型的用法，探讨该模型的适用性及
可行性，也具有一定的理论意义。

### 现实意义

一方面，从模型使用角度来看，对于空气质量的研究，国内大部分学者都是直接基于空气质
量指数通过建立回归方程进行研究，也有部分学者是以空气质量达到二级的天数为因变量进
行分析，但其建立的模型依然是普通回归模型。本项目除以空气质量指数为因变量外，还从
其他角度出发，分别以35个城市空气质量为优的天数和是否出现严重污染作为因变量，考虑
气象及经济因素对空气质量为优的天数及是否出现严重污染的线性及非线性影响，建立广义
可加混合模型对其进行研究。GAMM模型突破了传统回归模型模型设定的限制，加入了随机效
应，同时可以很好的反映变量间线性、非线性关系的影响。在空气质量影响因素的研究中引
入GAMM模型，探究其适用性和可行性，模型设定可能更为精确。

另一方面，就选题内容来说，清洁的空气是人类健康和福祉的基本需求，健康必须是城市规
划者的首要重点，空气质量影响着人们的生活，空气污染是影响健康的一个主要环境风险。
目前，空气污染非常严重，在全球范围内对健康构成了严重的威胁，PM2.5这个词汇经常会
出现在我们的生活中。《世卫组织空气质量准则》对空气污染的健康影响和有害健康的污染
阈值进行了评估。最新数据表明，全球80%以上的城市超过了世卫组织制定的空气质量限值，
在检测空气污染的所有城市中，一半以上报告的污染水平比世卫组织的限值高3.5倍或更多。
空气污染是一个隐秘杀手，中国是空气污染最严重的地区之一，对中国35个重点城市年度和
月度的空气质量进行分析，研究空气质量的影响因素，可以证明针对中国的实际情况，气象
因素和经济因素对空气质量的具体影响效应，并及时为短期内空气质量情况进行较为准确的
简单预报，对空气质量的改善有极大的推动作用。

## 国内外研究现状

关于广义可加混合模型的研究主要集中在医学、生物学、社会经济等领域，该模型可以很好的
处理各种类型的自变量，且引入了随机效应来反映不同对象之间的异质性以及同一对象不同
观测之间的相关性，综合了参数、非参数及随机效应的影响。但GAMM模型面临的主要问题是
参数估计问题，GAMM模型中有多个类型的参数，且每个参数还可能是多维的，传统的似然函数
估计方法很难实现。近年来，贝叶斯学派的统计推断方法受到了一些学者的广泛关注，贝叶斯
框架下后验分布的计算难免要涉及到高维数值积分，MCMC抽样方法的出现逐渐成为在统计推断
中解决高维数值积分问题强大而普遍的工具。一些学者也陆续提出在一些统计软件中基于不同
抽样方法的MCMC估计的实现程序，用于求得所建模型的估计结果。

### 国外研究现状

关于广义可加混合模型的发展，Nelder等（1972）提出了广义线性模型，并给出了基于似然
的参数估计的过程[@Nelder1972]。Hastie等（1986）通过光滑函数的和代替广义线性模型
中的线性形式提出广义可加模型，并介绍了二元因变量和生存数据的具体情况
[@Hastie1986]。 如同线性回归模型可以推广到广义线性模型一样，线性混合模型也可以推
广到广义线性混合模型(GLMM)。Laird（1982）提出了基于纵向数据的线性混合模型
[@Laird1982]，Gilmour（1985）针对二项分布因变量建立了广义线性混合模型
[@Gilmour1985]。Breslow和Clayton提出的广义线性混合模型(GLMM)为各种过度离散和相关
因变量的参数回归提供了统一的似然框架 [@Breslow1993]。GLMM一个关键特征是其使用参
数均值函数来对自变量的效应进行建模，并通过向线性预测子添加随机效应来拟合过离散化
和相关性。然后，这个参数均值函数的假设不一定总是正确的，因为合适的自变量函数形式
事先并不知道，因变量和自变量之间的依赖关系有可能具有复杂的形式。因此，向GLMM模型
中添加非参数均值函数项，建立相关数据的非参数回归模型具有实际意义。这将允许因变量
和自变量之间具有更灵活的函数依赖关系。GLMM推断的主要困难在于完全似然分析中难以解
决多重数值积分问题。各种近似推断方法(Breslow和Clayton; Lee和Nelder; Lin和
Breslow) [@Breslow1993;@Lee1996;@lin1996bias]，基于EM算法和Gibbs抽样的贝叶斯方法
(McCulloch; Zeger和Karim)可以用于解决这个问题
[@mcculloch1997maximum;@zeger1991generalized]。在1995年，Breslow等人提出GLMM中的
PQL估计是有偏的，其估计精度是值得怀疑的[@Breslow1995]。

对于独立数据，使用核或者样条等方法的非参数回归文献很多(Hardle; Green和
Silverman)[@hardle1990applied;@green1993nonparametric]。Hastie和Tibshirani提出的
广义可加模型应用非常广泛[@hastie1990generalized]。但是对于相关数据非参数回归的研
究却非常有限。大部分研究都局限于具有正态因变量的纵向数据并带有一个非参数函数
(Hart; Rice和Silverman)[@hart1994automated;@rice1991estimating]。还有一些研究在
线性混合模型中加入一个非参数时间函数(Zeger和Diggle; Zhang 等; Diggle和
Verbyla)[@zeger1994semiparametric;@Zhang1998;@verbyla2002analysis]。对于非正态纵
向数据，Wild和Yee与Berhane和Tibshirani把广义可加模型推广到广义估计方程
[@yee1996vector;@berhane2008generalized]。而在混合效应模型框架内研究非正态相关自
变量的非参数回归文献不多，具体可以参见Verbyla关于独立非正态数据广义线性模型中光
滑样条的混合模型公式的讨论[@verbyla2002analysis]。

相关数据的非参数回归面临许多新的问题。除了必须开发非参数函数的估计程序，我们还必
须考虑如何估计相关参数。许多学者(Green和Silverman; Wahba)强调过的一个重要问题是
如何选择好的光滑参数和带宽参数[@green1993nonparametric;@wahba1990spline]。这些问
题的相关研究较少，特别是相关参数和光滑参数的估计。传统光滑参数估计方法，数据驱动
方法面临着许多新的问题。例如，尽管交叉验证方法对于相关数据的光滑参数选择是一个合
理的方法，但其通常计算强度很大，而且接下来相关参数估计非常困难(Zeger和
Diggle)[@zeger1994semiparametric]。此外，交叉验证方法对于交叉设计和空间数据无效。
因此，开发同时对模型所有参数进行估计的系统方法具有重要意义。

广义可加混合模型是广义线性混合模型的推广，Lin和Zhang（1999）提出了广义可加混合模
型，在GLMM中引入非参数均值函数来研究非参数回归，可以更好的解决因变量和自变量之间
复杂的函数形式 。这类新模型用可加非参数函数建立自变量效应，在可加的预测子中加入
随机效应解释过离散和相关的问题。同时，他们也提出参数估计方法，用光滑样条估计非参
数函数，用边际拟似然联合估计光滑参数和方差成分。边际拟似然是限制极大似然（Zhang、
Lin 和 Raz等，1998）[@Zhang1998]的推广，考虑到数值积分的问题，利用双惩罚拟似然
（DPQL）进行近似推断[@Lin1999]。MCMC方法可以用来解决贝叶斯推断中高维数值积分的问
题，进而求得所有感兴趣后验的样本，得到后验均值、中位数、分位置信带和预测密度的估
计，不必基于渐近正态性的猜想进行近似，也不需要任何固定程序。

Lin和Zhang提出广义可加混合模型(GAMM)来解决这些问题[@Lin1999]。GAMM是根据Hastie和
Tibshirani的思想对GLMM的可加模型推广。这种新类型模型使用可加非参数函数来拟合自变
量的效应，而通过对可加预测子上增加随机效应来应对数据中的过离散化和相关性。Lin和
Zhang使用光滑样条来估计非参数函数，使用边际拟似然同时给出光滑参数和方差成分的估
计。边际拟似然方法是限制最大似然(REML)方法的推广[@Lin1999]。REML方法被Wahba和
Kohn用在经典非参数回归模型中[@wahba1985comparison;@kohn1991performance]，被Zhang
等，Brumback和Rice和Wang用在正态非参数混合模型中
[@brumback1998smoothing;@wang1998smoothing]，这些学者把光滑参数看作是额外的方差
成分。因为在最大化目标函数时需要求数值积分，双重惩罚拟似然(DPQL)被用来作近似推断。
上面方法的一个关键特征是在统一的参数混合模型框架下对GAMM的所有部分同时进行了系统
推断。具体地讲，GAMM的非参数函数、光滑参数和方差成分的估计可以通过已有的统计软件
拟合一个GLMM来得到。当数据是稀疏的时候(例如二分类数据)，方差成分的DPQL估计具有一
定的偏倚，这个偏倚可以通过一定的方法修正以改进这个估计。关于广义可加混合模型的估
计和推断更详细的讨论可以参考Lin，Zhang和Wood2017年的著作[@Lin1999;@Wood2017]。
Fahrmeir、Lang（2001）提出了通过基于Gibbs抽样的MCMC进行GAMM推断的一般贝叶斯方法
[@Fahrmeir2001]。鉴于MCMC中比较常用的抽样方法MH抽样和Gibbs抽样的局限性，
Neal（2003）提出切片抽样，并验证了其有效性[@Neal2003a]。Pham 和Wand(2018)在Zhao
等(2006)的研究基础上，对广义可加混合模型实现了一种新的基于切片抽样的MCMC参数估计
方法[@pham2018generalised;@Zhao2006]。


广义可加混合模型软件实现方面，在统计软件R中，有很多包支持GAMM的分析。比如，
Scheipl（2011）介绍了R包spikeSlabGAM的使用，可以在贝叶斯框架下实现GAMM模型的变量
及模型选择和正则化[@scheipl2011spikeslabgam]。Umlauf等（2015）介绍了R包R2BayesX
的使用，可以通过灵活的MCMC抽样对GAMM进行拟合[@umlauf2015structured]。Wood（2017）
[@Wood2017]介绍了R包mgcv的使用，通过gamm函数拟合GAMM模型，gamm给出了所有随机效应
和相关结构，可以具体化造成数值问题或PQL迭代失败的模型。Wood、Scheipl（2017）
[@Wood2017a]介绍了R包gamm4的使用，通过lme4函数实现GAMM的拟合，提出gamm4在数值上
比gamm更加稳健，并且避免了PQL估计对二元变量及低均值计数数据的过度拟合。Pham和
Wand（2018）阐述了R包gammSlice的使用，用于GAMM的基于切片抽样的贝叶斯拟合和推断，
发现与已经存在的统计软件中的函数相比，gammslice改善了推断精度
[@pham2018generalised]。在以上R包中，mgcv和gamm4是使用拉普拉斯近似的PQL方法实现
GAMM的参数估计，spikeSlabGAM、R2BayesX和gammSlice是通过MCMC进行拟合。

国外关于空气质量的研究也比较普遍，比如，Zanin、Marra（2012）利用可加混合模型研究
了澳大利亚、奥地利、加拿大、丹麦、芬兰、法国、意大利、西班牙和瑞士9个国家的环境
库兹涅茨曲线，发现经典的库兹涅茨曲线只适用于法国和瑞士，而澳大利亚、奥地利和西班
牙的人均二氧化碳排放与人均国内生产总值呈非线性增长[@zanin2012assessing]。Zheng等
（2017）使用中国2003-2012年的83个城市的数据，以当年PM10的水平为因变量，考虑人均
收入及其平方对PM10的影响，更系统地研究了中国的环境库兹涅茨曲线[@zheng2017a]。
Aung等（2018）对缅甸仰光室内室外空气质量进行了初步评估，发现室外空气质量被污染物
污染，室内空气质量随建筑物的特性而变化[@aung2018preliminary]。Han等（2019）使用
中国2016年AQI及相关气象及经济数据分别从全国和各城市的角度，建立模型对空气质量指
数的影响因素进行分析[@han2019analysis]。Zhu等（2019）根据气候条件变化及生物和人
口排放的变化，考虑各种情形全面评估了2035年三个主要加州空气流域的未来空气质量的主
要驱动因素[@zhu2019comprehensively]。

### 国内研究现状

对于模型的运用，国内大部分学者还是集中于广义线性混合模型及广义可加模型，关于广义
可加混合模型的应用相对较少，但也有学者涉及。比如，张继磊（2009）通过构建伽玛可加
混合模型、泊松可加混合模型，将非参数平滑方法运用到广义可加混合模型中，结合贝叶斯
理论和MCMC方法得到参数的估计，研究了GAMM在非寿险公司未决赔款准备金评估中的运用，
发现与传统的模型相比，GAMM得到的结果更为准确、更为客观[@张继磊2009]。关于贝叶斯
估计，很多学者都讨论了不同模型下的贝叶斯估计，包括多元线性回归模型（刘鹤飞，2019）
[@刘鹤飞2019]，二元选择分位数回归模型（邸俊鹏、张晓峒，2019）[@邸俊鹏2019]，空间
混合效应随机前沿模型（蒋青嬗、韩兆洲，2018）[@蒋青嬗2018]，非参数回归模型（苏雅
铃、何幼桦，2018）等[@苏雅玲2018]。但是还未曾有人单独对GAMM模型的估计方法进行说
明。针对本项目所涉及的参数估计方法，曹红艳、刘桂芬、曾平等（2008）比较了PQL和基
于MCMC的贝叶斯方法在广义线性混合模型中参数估计的精度和偏差[@曹红艳2008]。对于
MCMC方法中的切片抽样，陈小驽（2007）及王耀富（2008）通过指出常用的MCMC方法的不足，
引入切片抽样方法，通过模拟分析得出切片抽样可以降低计算成本的结论[@陈小驽2007;@王耀富2008]。

关于空气质量影响因素的研究方法，研究者大多采用环境库兹涅茨曲线（李健、靳泽凡、苑
清敏，2019），面板回归分析（王斌会、王术，2015），空间分析（蔺雪芹、王岱，2016），
灰色关联分析（湛社霞、匡耀求、阮柱，2018），主成分分析（李经路、曾天，2017）等
[@李健2019;@王斌会2015;@蔺雪芹2016;@湛社霞2018;@李经路2017]。在进行空气质量分析时，
主要采取以下两种变量来表征空气质量状况。第一类指标是按环境空气质量标准测度的环境
空气质量指数（AQI）（蔺雪芹、王岱，2016）以及由PM2.5、PM10、一氧化碳、二氧化硫、
二氧化氮、臭氧等为空气质量评价指标得出的空气质量综合指数或者直接根据主要污染物比
如PM2.5、PM10等为因变量进行研究（李茜、宋金平、张建辉，2013）[@蔺雪芹2016;@李茜2013]。
第二类是以某个时段内空气质量达标天数反映空气质量状况（吴雪萍、高明、曾岚
婷，2017）[@吴雪萍2018]。同时，在考虑空气质量的影响因素时，主要考虑两方面的因素。
第一类是社会经济因素，主要表现在人口因素、经济发展、技术水平、交通运输、绿化建设、
能源消耗等方面。第二类是气象因素，主要包括降水量、风速、气压、相对湿度、日照时数
等。其中，部分学者只考虑了社会经济因素（曲长雨，2018）[@曲长雨2018]，部分学者只
考虑了气象因素（李琛、刘瑾、王彦民，2017）[@李琛2017]，也有部分学者综合考虑了经
济和气象因素（李静萍、周景博，2017）[@李静萍2017]。

关于空气质量的研究范围，部分学者对全国空气质量进行了研究，比如，菜怡静等（2015）
运用2014年1月1日到4月30日161个城市的日AQI数据，李莉娜等（2017）运用2006-2012中国
各省的空气质量及经济社会和气象的面板数据，研究了城市空气质量的影响因素 [@蔡怡静2015;@李莉娜2017]。
也有部分学者只关注于某一个或某些区域，比如，周景博等（2015）
研究了北京市污染指数及其影响因素[@周景博2015]，徐冰烨等（2017）研究了2005-2014年
浙江省城市环境空气质量变化趋势及影响因素[@徐冰烨2017]。就全国范围来看，菜怡静等
（2015）以161个城市的日均空气质量指数作为因变量建立回归模型，发现常住人口、房屋
建筑施工面积、工业二氧化硫及烟尘排放总量、民用汽车拥有量对空气质量均呈正向影响，
绿化面积、3级以上风天数比例、降雨天数比例、平均气温对空气质量均呈负向影响 
[@蔡怡静2015]。王斌会等（2015）以中国31个城市为研究对象，以一年中空气质量达到或好于二
级的天数为因变量建立面板回归模型进行研究，发现PM10与城市空气质量之间存在显著负相
关关系，SO2、NO2、城市绿化的发展对城市空气质量的影响不显著，经济发展和公共交通发
展与城市空气质量大部分呈正比[@王斌会2015]。曲长雨（2018）以35个城市为研究对象，
以空气质量指数为因变量建立面板回归模型，得出人均地区生产总值、人口密度、单位GDP
用电量、公路客运量对空气质量指数存在正向影响，而城市绿化覆盖率对空气质量指数呈负
向影响的结论[@曲长雨2018]。李莉娜等（2017）以中国31个省为研究对象，以空气质量综
合指数为因变量构建环境库兹涅茨曲线，发现空气质量与人均国内生产总值、人口密度呈非
线性关系，空气质量综合指数与第一产业占比显著负相关、与第二、第三产业占比没有显著
的相关关系，与单位面积能源消耗、单位面积煤炭消耗显著正相关，与降水量呈显著负相关
关系[@李莉娜2017]。综合以上研究结果发现，不同学者对空气质量影响因素的研究得出的
结论不同，目前还没有统一的标准去判断经济社会及气象因素对空气质量影响的作用。

<!-- [@范永辉2007;@谢俊2009;@傅军和2008;@曹晨2007;@朱彬彬2018;@谷恒明2018;@刘霞2018;@汪卓蓉2018;@杨阳2016;@贾彬2005;@孙维伟2014;@孙煜2018;@茆诗松1999] -->

## 本项目的研究思路、主要内容和研究方法

### 研究思路

本项目的主要思路是首先通过线性混合模型和广义线性混合模型引入广义可加混合模型，并
阐述PQL估计及基于切片抽样MCMC方法的相关理论。进而选取中国按新标准监测并发布环境
空气质量的35个城市2014-2018年的空气质量年度和月度数据，利用数据对城市空气质量状
况进行基本描述并分析了其时空分布特征，并从社会经济因素、气象因素、地理因素分析了
城市空气质量影响因素的基本概况，进而对各城市年均空气质量指数及年均空气质量为优的
天数进行年度变化分析。

选取相关的经济社会因素及气象因素建立GAMM模型，综合考虑模型拟合优度，模型实际意义
等因素选择最优模型，并对最优模型给出合理的解释。一般来讲，空气质量会受到气象因素
和经济因素的影响，且根据以往的研究结果发现，有些因素对空气质量是线性影响，有些因
素对空气质量是非线性影响，且现有的研究大部分都直接以空气质量指数（air quality
index，AQI）为因变量。基于此，本项目除以空气质量指数为因变量外，还从其他角度出发，
分别以35个城市空气质量为优的天数和是否出现严重污染作为因变量，分别以年度空气质量
为优天数，月度空气质量为优天数，年度空气严重污染天数和月度是否出现空气严重污染为
考查变量，根据数据结构建立广义可加混合模型进行相应研究。

分别对年均AQI、年度空气质量为优的天数、年度严重污染的天数、月均AQI、月度空气质量
为优的天数、月度是否存在严重污染进行建模，通过这些因变量来分析空气质量的影响因素。
根据因变量的具体分布，且遵循模型由简单到复杂的原则，对服从正态分布的年均AQI、对
数化后的年度空气质量为优的天数及对数化后的月均AQI建立普通线性回归模型，进一步分
别考虑空气质量城市之间的相关性及自变量对空气质量的非线性影响，分别建立线性混合模
型及半参数混合模型；对于服从泊松分布的年度严重污染的天数、月度空气质量为优的天数，
建立泊松回归模型，同样分别考虑城市之间的相关性及自变量的非线性影响，分别建立泊松
混合模型及泊松可加模型，最后综合考虑城市之间的相关性及自变量的非线性影响，建立泊
松可加混合模型进行相应分析。同样，对于服从二点分布的月度是否发生严重污染，分别建
立logistic回归模型、Logistic混合模型、Logistic可加模型及Logistic可加混合模型进行
相应分析。最后在这些可选模型中通过综合比较模型拟合优度，模型实际意义等确定最优模
型。

### 主要内容

本项目的基本内容如下：

第一部分，导论。提出研究背景及选题意义，梳理国内外关于GAMM模型、MCMC方法、切片抽样、
PQL估计及空气质量影响因素的有关文献，明确研究目的。

第二部分，广义可加混合模型的模型背景和模型设定。在给出广义线性模型，线性混合模型
和广义可加模型等模型基础上，给出了广义可加混合模型的模型设定，特别是给出了贝叶斯
广义可加混合模型的模型设定。

第三部分为GAMM的参数估计方法，第一种方法为PQL估计，第二种方法为基于切片抽样的
MCMC估计。在PQL估计中，简述PQL估计的过程及步骤，在MCMC方法下，简介MCMC相关理论及
切片抽样，写出基于切片抽样MCMC估计的具体过程及收敛性诊断方法。

第四部分，城市空气质量状况描述及时空分布特征。首先进行变量选择，样本范围定义为重
点监测的35个城市，选取空气质量及其影响因素的2014-2017年的月度数据，主要包括空气
质量指数，空气质量等级的天数，考虑经济社会因素，气象因素及地理因素。经济社会因素
考虑经济发展、能源消耗、交通运输、人口密度；选定的指标有：规模以上工业增加值、固
定资产投资、全社会用电量、公路客运量、地区生产总值、人口密度、城市绿化覆盖率、二
氧化硫及烟尘排放总量等。选定的气象因素有：降水量、平均气温、日照时数、相对湿度。
选定的地理因素有：是否供暖、纬度、海拔、是否临海。经济社会因素中某些指标的月度数
据不容易获得，因此对于地区生产总值等季度数据会用合适的频率转化方法进行相应的频率
转化；针对时点数据比如城市绿化覆盖率，因其在一年范围内变动较小，因此直接将年末数
据当作本年每月的月度数据进行分析。之后进行城市空气质量状况描述；最后对空气质量指
数的时空分布特征进行分析，进而计算全局空间自相关指数，观察空气质量指数的时空格局
演变。

第五部分，对城市空气质量指数影响因素进行实证分析。首先进行因变量定义，之后进行变
量筛选，使用Lasso方法筛选出主要变量，再分别考虑各城市年度空气质量指数和月度空气
质量指数，建立广义可加混合模型的各种具体模型，通过模型选择获取最优模型，并对最优
模型进行解释，分别从众多可能因素中得到年度和月度空气质量指数的影响因素。

第六部分，从城市空气质量为优天数的角度进行实证分析。首先分别从年度和月度对城市空
气质量为优天数的分布进行分析，然后在Lasso变量选择的基础上对年度空气质量为优天数
建立正态可加混合模型的各种具体模型，对月度空气质量为优天数建立泊松可加混合模型的
各种具体模型，最后通过模型选择得到最优模型，并对最优模型进行解释。

第七部分，从城市空气质量严重污染的角度进行实证分析。对城市年度空气严重污染的天数
建立泊松可加混合模型的各种具体模型并选择最优模型。对城市月度是否会出现严重污染的
二分类变量建立Logistic可加混合模型并选择最优。对两个最优模型进行分析，从而得到城
市空气质量严重污染的影响因素。 

第八部分，结论。对本项目的研究内容进行总结，根据实证分析得出空气质量影响因素、具
体影响情况及影响程度大小的相关结论，最终指明本项目研究的不足及未来的研究方向。

### 研究方法

1. 文献研究法。查阅国内外相关文献和书籍，搞清楚GAMM的具体涵义及适用条件，弄明白
PQL估计和基于切片抽样MCMC估计的基本思想，整理关于空气质量的相关研究；

2. 统计方法。由于本项目是研究一类统计方法及其应用，故本项目中将综合应用大量的统计
方法。在实证分析中，除了使用广义可加混合模型之外，我们还将运用描述统计方法，线性
回归方法，线性混合模型方法，半参数回归方法等等。

3. 定性研究方法。尽管本项目的研究主要使用以统计方法为主的定量研究方法。但定量研究
方法不能完全取代定性研究方法。本项目拟在分析和解释统计模型结果时使用定性研究方法。

## 本项目可能的创新点

（1）就使用的模型来看，目前国内大部分学者在进行GAMM估计时基本上都是利用基于M-H抽
样和Gibbs抽样的MCMC方法，还未有人将基于切片抽样的MCMC估计方法应用于GAMM，本项目
应用H.Pham在2018年最新提出的R包gammSlice对广义可加混合模型进行更加精确的估计。

（2）就研究主题来看，通常对于空气质量的研究是基于31个省或部分城市的年度数据，用
月度数据研究的相对较少，且研究方法一般集中于环境库兹涅茨曲线、面板回归模型等，还
未涉及广义可加混合模型的应用。本项目选取2014-2017年中国重点监测的35个城市的月度
数据建立广义可加混合模型进行相应研究。此外，除把空气质量指数作为因变量分析外，还
借助于广义可加混合模型的广泛适用性，研究了空气质量为优天数和出现严重污染好坏两种
极端情况，使得对于空气质量的研究更完整。


<!-- # 参考文献 -->
[//]: # (\bibliography{Bibfile})
