==> report/04-aqi-des.Rmd <==
---
title: "城市空气质量状况描述及时空分布特征"
author: "Jin"
date: '2019-04-15'
csl: ./style/chinese-gb7714-2005-numeric.csl
css: ./style/markdown.css
bibliography: Bibfile.bib
eqnPrefixTemplate: ($$i$$)
link-citations: true
linkReferences: true
notice: '@*'
autoEqnLabels: true
---

```{r setup, echo=F, message=F}

################# 第 4 章 R 程序代码  ####################


knitr::opts_knit$set(root.dir = getwd())
knitr::opts_chunk$set(echo = FALSE, results = 'hide')
knitr::opts_chunk$set(warning = FALSE, message=FALSE)
options(knitr.kable.NA = '')
```

```{r prepare}
rm(list=ls())
options(digits=4)
options(scipen=100)
graphics.off()
Sys.setlocale("LC_ALL", "Chinese")
windowsFonts(msyh=windowsFont("微软雅黑"))
library("kableExtra")
```

# 城市空气质量状况描述及时空分布特征 {#sec:chap-4}

在第 \@ref(sec:chap-2) 章和第 \@ref(sec:chap-3) 章中，我们对广义可加混合模型的模
型背景，模型设定和参数估计方法进行了讨论。本章和下面三章转向城市空气质量的实证部
分。本章对空气质量指数进行定义，选取变量，收集数据，并对收集的数据进行预处理。利
用收集的数据对35城市空气质量现状进行描述，并分析了35城市空气质量指数的时间和空间
分布特征。

## 空气质量指数简介 {#sec:4-1}

### 空气质量指数定义和分级

在《环境空气质量标准》中，将环境空气定义为人群、植物、动物和建筑物所暴露的室外空气[@aqi]。环境空气
功能区分为两类：一类区为自然保护区、风景名胜区和其他需要特殊保护的区域；二类区为居住区、商业交
通居民混合区、文化区、工业区和农村地区。一类区适用一级浓度限值，二类区适用二级浓度限值。环境
空气常规污染物基本浓度限值如表 \@ref(tab:tab-air-grade) 所示。

```{r tab-air-grade, eval=T,results='markup', cache=F}
tab1 <- read.csv('./results/airpollution.csv')
knitr::kable(tab1, row.names =F, align = "l", caption="环境空气污染物基本项目浓度限值",
      longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F)%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header"),
                  stripe_index = c(1:3,7:8,11:12))%>%
    kable_styling(font_size = 11)%>%
    footnote(general ="生态环境部发布《环境空气质量指数（AQI）技术规定（试行）》",
             general_title = "数据来源：")
```

注：颗粒物（粒径小于等于10 $\mu m$）即PM10,指的是环境空气中空气动力学当量直径小于等于10
$\mu m$ 的颗粒物，也称可吸入颗粒物；颗粒物（粒径小于等于2.5 $\mu m$ ）即PM2.5指的是环境
空气中空气动力学当量直径小于等于2.5 $\mu m$ 的颗粒物，也称细颗粒物。1小时平均指
任何1小时污染物浓度的算术平均值；8小时平均指连续8小时平均浓度的算术平均值，也称8
小时滑动平均；24小时平均指一个自然日24小时平均浓度的算术平均值，也称为日平均；年
平均指一个日历年内各日平均浓度的算术平均值。

《环境空气质量指数（AQI）技术规定（试行）》对空气质量指数及常规污染物进行了定义。空气质量
指数（air quality index，AQI）是定量描述空气质量状况的无量纲指数，是空气质量分指数中的最大
值。空气质量分指数（individual air quality index，IAQI）指单项污染物的空气质量指数。用数学
语言表述即为： $\mathrm{AQI}=\max \left\{\mathrm{IAQI}_{1}, \mathrm{IAQI}_{2},
\ldots, \mathrm{IAQI}_{\mathrm{n}}\right\}$ ，
其中 $n$ 表示污染物项目。 AQI 大于50时,IAQI 最大的污染物为首要污染物，若
IAQI 最大的污染物为两项或两项以上，则并列为首要污染物。 IAQI 大于100的污染物为超标污染物。

空气质量指数相关信息如表 \@ref(tab:tab2) 所示。当空气质量为优时，空气质量令人满意，基本无空气污染，各类人群
可正常活动。当空气质量为良时，空气质量可接受，但某些污染物可能对极少数异常敏感人群健康有较
弱影响，此时，极少数异常敏感人群应减少户外活动。当空气质量为轻度污染时，易感人群症状有轻度
加剧，健康人群出现刺激症状，儿童、老年人及心脏病、呼吸系统疾病患者应减少长时间、高强度的户
外锻炼。当空气质量为中度污染时，会进一步加剧易感人群的症状，可能对健康人群心脏、呼吸系统有
影响，此时，儿童、老年人及心脏病、呼吸系统疾病患者应避免长时间、高强度的户外锻炼，一般人群
也要适量减少户外运动。当空气质量为重度污染时，心脏病和肺病患者症状显著加剧，运动耐受力降低，
健康人群普遍出现一些症状，这时，儿童、老年人、心脏病和肺病患者应停留在室内，停止户外运动，一般人
群也需要减少户外运动。当空气质量为严重污染时，健康人群运动耐受力降低，有明显强烈的症状，会
提前出现某些疾病，儿童、老年人和病人应当留在室内，避免体力消耗，一般人群也应避免户外活动。

```{r tab2, eval=T,results='markup', cache=F}
tab2 <- read.csv('./results/airindex.csv')
knitr::kable(tab2, row.names =F, align = "l", caption="空气质量指数及其等级划分",
             longtable = TRUE, booktabs = TRUE, linesep  = "")%>%
    kable_styling(full_width = T) %>%
    column_spec(1:3, width = "4cm")%>%
    footnote(general ="生态环境部发布《环境空气质量指数（AQI）技术规定（试行）》",
             general_title = "数据来源：")
```

2013年起，环境保护部要求对六种常规污染物进行监测，包括二氧化硫、二氧化氮、一氧化
碳、臭氧、PM10 和 PM2.5 。《2017中国生态环境状况公报》指出，2017 年，全国 338 个
地级及以上城市中，有 99 个城市环境空气质量达标，占全部城市数的 29.3% ，239 个城
市环境空气质量超标，占 70.7% 。338 个城市发生重度污染 2311 次、严重污染 802 次，
以 PM2.5 为首要污染物的天数占重度及以上污染天数的 74.2% ，以 PM10 为首要污染物的
占 20.4% ，以臭氧为首要污染物的占 5.9% 。有研究表明， PM2.5 及以下的微粒，75% 会
在肺泡沉积，长期停留在呼吸系统内，沉积在呼吸系统深处，人体对 PM2.5 没有任何过滤、
阻拦能力，而 PM2.5 对人体的危害极其恐怖。在欧盟国家中，PM2.5 导致人们的平均寿命
减少 8.6 个月，在众多污染物中，PM2.5 对健康的影响尤为显著。

### 空气质量影响因素概述

空气质量的好坏反映了空气污染的程度，空气污染是一个复杂的现象，在特定时间和地点，空气质量受
很多因素的影响，比如常见的空气质量影响因素包括经济社会因素、地理因素及气象因素。

#### 经济社会因素

现有的大部分研究表明，经济社会因素即人为因素是影响空气质量最重要的因素，主要表现在人
口因素、经济发展、技术水平、交通运输、绿化建设、能源消耗等方面。空气质量的好坏主要受
污染物排放的影响，汽车尾气、工厂废气、建筑施工、居民生活和取暖、垃圾焚烧等都会产生一
定程度的污染排放。一般来讲，一个地区经济发展越好，人口越多，交通越发达，能源消耗越多，
其污染物排放越多，相应地空气质量也越差；一个地区绿化建设越好，污染治理投入越多，其空
气质量会越好。

#### 地理因素

地理因素主要包括地形地貌、海拔、是否临海等。地形影响气流，气流影响空气流向。通常情况下，
地势高的地方空气流通较好，有利于污染物的扩散，空气质量相对较好。同样，地处海边的城市，
因为有海风和气流的影响，污染物更易扩散，所以相对内陆来讲，临海城市的空气质量一般都要比
内陆好。

#### 气象因素

气象因素是影响空气质量的主要因素之一。通常影响空气质量的气象因素包括气温、气压、
风向、风速、降水量、相对湿度等。大气污染在垂直方向的扩散主要取决于气温的垂直分布，
当气温较高时，大气处于不稳定状态，在热力对流的作用下污染物向上扩散，空气质量就会
变好；反之，当气温较低时，大气变得稳定，污染物的扩散受到抑制作用，空气质量就会变
差。一般来讲，当低压控制时，大气处于中性或不稳定状态，低层空气辐合上升，近地面的
污染物随空气上升到高空，有利于近地面污染物向高空扩散；当高压控制时，空气作下沉运
动，并常形成下沉逆温，阻止污染物的向上扩散。风是边界层内影响污染物稀释的重要因子，
风速是造成快速水平输送或平流的主要原因，而风向则决定大气污染物浓度的分布。风速对
污染物的浓度环境浓度具有双重影响，在一定范围内，风速越大越有利于空气污染物的扩散
和稀释，空气质量越优；超过这一范围，风速增大将使空气中可吸入颗粒物浓度明显增加，
导致空气污染加重。针对降水量，大气降水不仅可以冲刷空气中的部分颗粒物，也可以在一
定程度上抑制地面扬尘发生，从而有效控制颗粒物的排放，对空气质量有较好的净化作用。
相对湿度能够促进一次污染物向二次颗粒污染物转化，相对湿度较大时，由于常规污染物周
围被水分包裹，质量加重，重力增大造成一定程度沉降，从而减小污染物浓度，空气质量相
对较好。

## 变量选择及数据预处理

### 变量选择

空气质量的好坏反映了空气污染的程度，空气质量指数是反映空气质量好坏最重要的指标之一，
根据空气质量指数划分的空气质量等级可以对空气质量进行更加直观的判断。严重的空气污染会
导致健康人群运动耐受力降低，有明显强烈的症状，身体出现各种不适，对机体的健康产生一定
的影响，提前出现某些疾病。空气质量为优是我们所期待出现的结果，空气质量令人满意，基本
无空气污染，各类人群可正常活动。这就为我们在进行空气质量的研究时提供了一个全新的视角。
进而结合数据特征，我们可以选取月均空气质量指数、每月是否存在严重污染及月度空气质量为
优的天数作为因变量进行相应分析。

在自变量的选取过程中，通过查阅相关文献，结合空气质量的影响因素及数据的可获得性，可以
初步从经济社会因素、气象因素和地理因素中选取以下指标进行具体分析。经济社会因素中，经
济增长、人口因素、绿化建设、交通运输、能源消耗、废气排放等都会对空气质量产生影响。地
区生产总值是衡量一个地区经济发展的重要指标，固定资产投资在一定程度上也会拉动经济增长。
因此，初步选取地区生产总值和固定资产投资作为城市经济增长的指标。有研究表明，人口集聚
会对空气质量产生一定的影响，人口密度可以反映人口地理分布的疏密程度，但是考虑到市辖区
人口较多，郊区人口较少，从某种意义上讲，人口密度会将市辖区和郊区对空气质量的影响视为
等同，因此，选取年末常住人口表示人口因素。空气质量的两大污染源包括固定污染源和移动污
染源，固定污染源是指向环境排放或释放有害物质或对环境产生有害影响的场所、设备和装置，
用房屋建筑施工面积代表。移动污染源主要指空气排放污染物的交通工具，用私人汽车拥有量和
城市公路客运量进行表示。能源消耗产生的废气严重污染了空气，尤其是煤炭消耗，电能是清洁
高效的二次能源，用全社会用电量代表能源消耗。同时选取工业二氧化硫排放量和工业烟粉尘排
放量作为废气排放的指标。绿化建设可以吸附灰尘、净化空气、减少微粒并减少空气中细菌含量，
对空气的净化有重要的作用。同时，环境治理投资特别是废气治理力度的大小也会对空气质量的
好坏产生重要的影响。因此，选取建成区绿化覆盖率，工业污染治理完成投资和废气治理完成投
资作为绿化建设和污染治理的指标。

经济社会因素可以视为影响空气质量好坏的人为因素，这些因素可以根据实际情况做出相应改变。
而气象因素及地理因素则是影响空气质量好坏的固定因素，气温、降水量、气压、风速、相对湿度、
日照时数都会对空气质量产生一定的影响。同时，城市的地形、南北方差异，北方供暖会产生大量
废气排放，沿海城市因为有海风有利于污染物扩散都会对空气质量产生相应的影响。

根据以上分析，鉴于数据的可获得性，选取比较有代表性的指标，最终确定空气质量影响因
素一级指标为：社会经济因素、气象因素和地理因素。其中社会经济因素表现在经济增长、
人口因素、绿化建设、交通运输、固定污染源、移动污染源、能源消耗、废气排放、污染治
理这些方面。其中，固定资产投资、地区生产总值、地区生产总值增长率表示经济增长；常
住人口表示人口因素；建成区绿化覆盖率表示绿化建设；城市公路客运量表示交通运输；房
屋建筑施工面积表示固定污染源；私人汽车拥有量表示移动污染源；全社会用电量表示能源
消耗；工业二氧化硫排放量、工业烟粉尘排放量表示废气排放；工业污染治理完成投资和废
气治理完成投资表示污染治理。社会经济因素最后选定13个指标。最终选定的气象因素包括：
降水量、平均气压、平均风速、平均气温、平均相对湿度、日照时数等6个指标。最终选的
地理因素有：是否供暖、是否临海、经度、纬度、海拔，共5个指标。三个方面影响因素合
计选择了24个指标，最终的指标体系如表 \@ref(tab:tab3) 所示。

```{r tab3, eval=T,results='markup', cache=F}
library("kableExtra")

tab3 <- read.csv('./results/variable.csv')
knitr::kable(tab3, row.names =F, align = "l", caption="空气质量影响因素指标体系",
             longtable = TRUE, booktabs = TRUE, escape = F,
             linesep  = c(rep("", 12), "\\hline", rep("",5), "\\hline")) %>%
			     kable_styling(latex_options = c("scale_down", "repeat_header", "hold_position"),
                 repeat_header_text = "(续)")%>%
    kable_styling(full_width = T) %>%
    column_spec(1:4, width = c("3cm", "3cm", "5cm", "2cm"))
			 
```


### 数据来源及预处理

#### 数据来源

省会城市和计划单列市一向是我们研究的重点。在数据采集过程中，由于拉萨数据存在严重缺失，
因此本文研究范围为除拉萨以外的30个省会城市及5个计划单列市共35个城市。各城市基本情况
显示在表 \@ref(tab:tab4) 中。其中，城市基本数据中的经度、纬度、海拔来源于中国气象局发布的中国地面国际
交换站气候资料月值数据集，是否供暖根据国家公布的全国供暖时间表进行判断。

```{r tab4, eval=T,results='markup', cache=F}
tab4 <- read.csv('./results/city.csv')
knitr::kable(tab4, row.names =F, align = "l", caption="城市基本情况概述",
      longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F)%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)")%>%
    kable_styling(full_width = T) %>%
    column_spec(1:3, width = c("2cm", "2.5cm","2.5cm"))%>%
    footnote(general ="《中国气象年鉴2017》，其中纬度和经度是该城市气象观测站所在点的纬度和经度",
             general_title = "数据来源：")

```

由于空气质量是实时数据，年度数据分析可以了解基本概况，月度数据分析可能会使结果更
加准确，因此，本文从年度和月度两个时间频率对空气质量状况进行分析。考虑到数据的可
获得性，选取2014年1月到2017年12月的城市月度数据，相应的选取2014-2017年的年度数据
进行相应分析。在数据获取过程中，空气质量相关数据及气象指标可以直接获得月度数据，
其中，空气质量相关数据（空气质量指数、PM2.5、PM10、SO2、NO2、CO、O3）来源于中华
人民共和国生态环境部发布的2014年1月到2017年12月的《城市空气质量状况月报》，气象
相关数据（最低气温、最高气温、降水量、平均气压、平均风速、平均气温、平均相对湿度、
日照时数）来源于2015-2018年《中国统计年鉴》、2015-2017年《中国气象年鉴》及中国气
象局发布的中国地面国际交换站气候资料月值数据集。对于经济社会指标相关数据，在所涉
及的指标中，表示经济增长的地区生产总值季度数据及固定资产投资月度数据均来自于各城
市2014年2月到2017年12月的月度统计公报，数据可以从各城市的统计局官网获得。对于年
末常住人口、房屋建筑施工面积、城市公路客运量、私人汽车拥有量、全社会用电量、工业
二氧化硫、工业烟粉尘排放等年度数据，均来自2014-2018年相应的城市统计年鉴，对于部
分缺失数据，在各城市年度统计公报中查找。


#### 数据预处理

由于各城市月度固定资产投资数据均为累计值，非单月指标，因此，对于固定资产投资，采
用后一期减前一期的计算方法，得出本月值。同时由于1-2月为累计值，春节期间不单独对
单月数据进行统计，因此，本文采用取平均值的方法，用1-2月累计值的平均值作为1、2月
单月数据。

对于部分缺失数据，比如太原市2017年城市公路客运量，乌鲁木齐市2017年常住人口，由于
城市公路客运量和常住人口在不同年度变化幅度不是太大，因此，直接将2016年太原市城市
公路客运量和2016年乌鲁木齐市的常住人口作为2017年的数据进行填充。对于私人汽车拥有
量这一指标，比如呼和浩特市，已有的数据为2014年私人汽车拥有量和2014、2017年民用汽
车拥有量，用2014年私人汽车拥有量在民用汽车拥有量中的占比乘以2017年民用汽车拥有量
作为2017年私人汽车拥有量的数据，2014年私人汽车拥有量加2014-2017年的平均增长量作
为2015年的数据，同样2016年的数据也是在2014年数据的基础上加两年平均增长量。对于全
社会用电量这一指标，比如昆明市，已有2014、2017年的数据，用2014、2017年数据的平均
值作为2015年和2016年的数据。其他城市这些指标的缺失值均采用类似的方法进行处理。对
于气象因素中缺失数据的处理，均采用上年同期数据对本年当期数据进行填充。

对于工业污染治理完成投资和废气污染治理投资，各城市并未公布具体数据，但其对空气质
量的好坏有很大程度的影响。我们可以从中国统计年鉴中获得各城市所在各省份的工业污染
治理完成投资和废气污染治理投资数据。我们假定各城市工业污染治理完成投资和废气污染
治理投资占全省的比例与其GDP占全省的比例大致相等。因此，将各城市地区生产总值占所
在省份生产总值的比重乘以所在省份工业污染治理完成投资和废气污染治理投资作为各城市
这两个变量的数据。

由于要进行月度分析，在所获得的数据中，针对非月度数据，需要采用合适的频率转化方法
进行频率转化，转为月度数据进行分析。经济指标的数据类型可以分为存量与流量数据，对
于地区生产总值、城市公路客运量、房屋建筑施工面积、全社会用电量、工业二氧化硫排放
量、工业烟粉尘排放量、废气污染治理投资这些流量数据，季度地区生产总值除以3作为月
度地区生产总值的近似，其他年度数据均采用年度数据除以12的方法将数据平均分配到各月
中。对于常住人口、建成区绿化覆盖率、私人汽车拥有量这些存量数据，均将年末值作为各
月值进行分析。

## 35城市空气质量状况描述

### 城市空气质量状况描述

根据空气质量等级划分标准，当空气质量指数为0-50时，空气质量为优；当空气质量指数
在51-100时，空气质量为良；当空气质量指数在101-150时，空气质量为轻度污染；空气
质量指数为151-200时，空气质量为中度污染；当空气质量指数为201-300时，空气质量为
重度污染；当空气质量指数大于300时，为严重污染。清洁的空气是人们生产生活的基础，
空气污染会对人们的身体和生活产生负面影响，影响人们的健康和正常活动。为了直观来
看所研究的35个城市的空气质量状况，我们选取2017年的空气质量相关数据，计算了各城
市2017年年均AQI，并根据空气质量等级划分标准划分了各城市2017年的空气质量等级。
同时为了观察各城市2017年空气质量最好和最差的情况，下面给出2017年35个城市年均AQI、
空气质量等级为优的天数及其在当年所有天数中所占百分比和空气质量为严重污染的天数，
具体结果如表 \@ref(tab:tab-aqi-year) 所示。

```{r tab-aqi-year, eval=T, results='markup'}
dat.y <- read.csv("./data/yearly.csv")

aqi.y17 <- dat.y[dat.y$year==2017, c(1,3,10,15)]
aqi.y17$grade <- cut(aqi.y17$AQI, c(0,50,100,150), labels = c("优","良","轻度污染"))
aqi.y17$excellent.ratio <- aqi.y17$excellent/365*100
aqi.y17 <- aqi.y17[,c(1:2,5,3,6,4)]
names(aqi.y17) <- c("城市","年均AQI","等级","等级为优天数","等级为优百分比","严重污染天数")
write.csv(aqi.y17, "./results/aqi.y17.csv", row.names=F)

library("kableExtra")
knitr::kable(aqi.y17, row.names =F, align = "c", caption="2017年35个城市空气质量概况",
             longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F) %>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)")%>%
    footnote(general ="通过生态环境部城市空气质量状况月报数据整理计算得到",
             general_title = "数据来源：")

```

根据表 \@ref(tab:tab-aqi-year) 可知，在2017年，海口是唯一一个年均AQI小于50的城市，其年均空气质量为优
的天数达到168天，占全年总天数的46.027%，这就表明2017年，海口有一半左右的时间空气
质量都为优，空气质量很好，且在我们所研究城市中其空气质量最好，当然这与海口的气象
条件和地理位置也比较相关。相反，北京、天津、石家庄、太原、济南、西安、兰州、乌鲁
木齐年均AQI在100-150之间，处于轻度污染的状态，其余城市年均空气质量等级均为良。
同时可以发现，贵阳、昆明、厦门这三个城市年均空气质量为优的天数大于150天，在全年
占比大于40%。而石家庄、济南、兰州、银川在2017年全年空气质量为优的天数分别为6、7、
8、9天，其空气质量为优的天数在全年中占比小于3%。表明贵阳、昆明、厦门这些城市的空气
质量在所有城市中相对较好，而石家庄、济南、兰州、银川这四个城市的空气质量相对较差。
在2017年全年，石家庄严重污染的天数达到了18天，西安、乌鲁木齐严重污染的天数达到了15天，
且结合原始数据可以发现，石家庄的空气质量往往表现的都比较差，这也许是因为石家庄作为
我国北方的内陆城市，西邻太行山，且盛行北温带亚湿润大陆季风气候，当工业大气污染物随
东南风飘到太行山脉时，无法翻过山而停留在石家庄无法稀释，进而使空气中污染物浓度过高，
造成环境污染。且部分工业发展迅速，化学工业、煤炭工业亦占有重要地位，这些工业废气也
导致大气污染日益严重且难以治理。然而，像沈阳、大连、上海、杭州、宁波、合肥、福州、
厦门、南昌、青岛、广州、深圳、南宁、海口、重庆、成都、贵阳、昆明这些城市全年无严重
污染天气，这与其地理位置也有一定的关系。比如，像大连、厦门、青岛、深圳、海口这些沿海
城市，空气流通好，且沿海城市由于靠海，海水也会吸附一部分空气中的杂质，具有一定的环境
净化作用。再如，像贵阳、昆明这些空气质量较好的城市，除了受到政府政策的影响外，其均
处于云贵高原，虽然两者的气候不同，但其地理位置对空气质量的好坏还是具有很大的影响。

### 城市空气质量影响因素描述性分析

在对城市空气质量基本状况进行描述之后，进一步结合指标体系对城市空气质量影响因素进行
描述性分析，直观判断空气质量与各影响因素之间的关系。下面分别从社会经济因素、气象
因素、地理因素三个方面进行说明。

#### 社会经济因素变量描述性分析

根据已有的研究结果可以发现，社会经济因素是影响空气质量好坏的最重要的因素之一。为了
对所有城市的整体社会经济发展及其之间的差异做出直观判断，根据本文所建立的指标体系，
对2017年35个城市的社会经济统计变量进行描述性分析，计算各指标的均值、最小值、最大值
及相应的标准差。计算结果如表 \@ref(tab:tab-eco) 所示。

```{r tab-eco, eval=T, results='markup'}

varlist <- read.csv("./data/var_list.csv")
dat.y <- read.csv("./data/yearly.csv")
ecovar <- names(dat.y)[c(1, 17:29)]
eco.y17 <- dat.y[dat.y$year==2017, ecovar]

varinfo <- varlist[match(ecovar[-1], as.character(varlist$var_code)), c(1,2,7)]

eco.stat <- as.data.frame(t(apply(eco.y17[,-1], 2, function(x) c(mean(x), min(x), max(x), sd(x)))))
eco.stat <- cbind(varinfo, eco.stat)
eco.stat <- eco.stat[order(eco.stat[,1]),][,-1]

names(eco.stat) <- c("变量", "单位", "均值", "最小值", "最大值", "标准差")
write.csv(eco.stat, "./results/eco.stat.csv", row.names=F)

library("kableExtra")
knitr::kable(eco.stat, row.names =F, align = c("l","l",rep("r", 4)),
             digits=2,
             caption="2017年35个城市社会经济统计变量描述统计",
             longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F) %>%
    kable_styling(full_width = T) %>%
    column_spec(1:2, width = c("4.3cm","1.9cm"))%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  ## stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)",
                  )%>%
    footnote(general ="2018年《中国统计年鉴》，各城市统计年鉴的数据整理计算得到。",
             general_title = "数据来源：")

```

根据表 \@ref(tab:tab-eco) 可以发现，2017年35所城市的地区生产总值及固定资产投资
额均值分别为9465.48亿元、5277.06亿元，其中地区生产总值最小的是西宁，为1284.91
亿元，最大的是上海，为30632.99亿元；固定资产投资额最大的是重庆，为17537.05亿元，
最小的是海口为489.3亿元。地区生产总值是衡量一个地区经济发展最重要的指标之一，由
以上结果可以直观看出上海的经济发展最好，其地区生产总值大概是35所城市经济发展平均水平
的3.24倍，而西宁的经济发展最差，只有35所城市平均发展水平的13.57%。同时结合
空气质量相关数据，可以发现，2017年上海年均空气质量为良，无严重污染天气，而
西宁的年均空气质量也为良，但其有4天严重污染。理论上我们认为，经济发展越好的
城市其空气质量越差，但从数据中可以看出，上海经济发展比西宁经济发展好很多，但
其空气质量相差并不大，直观来看，社会经济因素并不是决定空气质量好坏的唯一原因。

除了表示经济增长的因素之外，人口因素、移动污染源等也是空气质量好坏的重要影响因素。
2017年35所城市人口最多的是重庆，为3075万人，人口最少的是银川，为222.54万人。城市
公路客运量最多的是广州，为91323万人，最少的是呼和浩特，为435万人。私人汽车拥有量
最多的是北京，为466.61万辆，最少的是西宁，为33.59万辆。全社会用电量最多的是北京，
为1066.89万千瓦时，最少的是海口，为70.34万千瓦时。工业二氧化硫排放量最多的是重庆，
为139880吨，工业烟粉尘排放量最多的是呼和浩特，为111036吨，两者排放最少的城市均为
海口，分别为502吨和113吨。作为固定污染源的代表，房屋建筑施工面积最多的是上海，为
41197.5万平方米，最少的是南宁，为1847.32万平方米。且对于污染的各种治理，建成区绿
化覆盖率最多的是北京，为48.42%，最少的是兰州，为30.93%。35个城市2017年年均工业污
染治理完成投资为7.06亿元，各个城市工业污染治理完成投资相差较大，标准差达到7.70亿
元；废气治理完成投资平均值为4.32亿元，其中废气治理完成投资额最多的是上海，为
13.35亿元，最少的是西宁，为0.29亿元。

根据以上信息可以发现，重庆的固定资产投资最多，人口最多，工业二氧化硫排放也最多，其
建成区绿化面积为40.32%，废气污染治理投资额为5亿元，2017年空气质量等级为良，无严重
污染天气。表明虽然重庆污染物排放比较多，但及时采取有效的环境保护措施使得其空气质量
并没有表现太差。同时，北京的私人汽车拥有量和全社会用电量在35个城市中最多，虽然其建成区
绿化覆盖率也是最多，废气污染治理投资额为7.85亿元，远高于各城市年均水平，但是其年均空气
质量等级为轻度污染。结合原始数据可以发现，35个城市中，北京在各种污染源及经济发展指标
中排名都比较靠前，这或许是由于北京是政治、经济中心决定的。房屋建筑施工面积最多的是
上海，废气污染治理投资最多的也是上海，上海的经济发展比较靠前，各种污染物排放也比较多，
但是其废气治理投资额也相对较多，因此，其2017年年均空气质量等级为良或许是因为其大力度
的污染治理。相反，海口的空气质量等级为优也可以从以上数据明显看出，海口的全社会用电量
最少，工业二氧化硫和烟粉尘排放也最少，这表明其污染源排放较少，其空气质量为优也得益于
这些条件。总的来说，污染物排放和废气治理对空气质量的好坏具有复杂的交互作用。

#### 气象因素变量描述性分析

与此同时，为了观察各城市气象因素的总体概况，进而对各城市的平均气温、降水量、平均相对湿度、平均风速、平均气压及日照时数进行描述性分析，同样计算各
指标的均值、最小值、最大值及相应的标准差。计算结果如表 \@ref(tab:tab-wea) 所示。

```{r tab-wea, eval=T, results='markup'}

varlist <- read.csv("./data/var_list.csv")
dat.y <- read.csv("./data/yearly.csv")
weavar <- names(dat.y)[c(1, 33:38)]
wea.y17 <- dat.y[dat.y$year==2017, weavar]

varinfo <- varlist[match(weavar[-1], as.character(varlist$var_code)), c(1,2,7)]

wea.stat <- as.data.frame(t(apply(wea.y17[,-1], 2, function(x) c(mean(x), min(x), max(x), sd(x)))))
wea.stat <- cbind(varinfo, wea.stat)
wea.stat <- wea.stat[order(wea.stat[,1]),][,-1]

names(wea.stat) <- c("变量", "单位", "均值", "最小值", "最大值", "标准差")
write.csv(wea.stat, "./results/wea.stat.csv", row.names=F)

library("kableExtra")
knitr::kable(wea.stat, row.names =F, align = c("l","l",rep("r", 4)),
             digits=2,
             caption="2017年35个城市气象变量描述统计",
             longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F) %>%
    kable_styling(full_width = T) %>%
    column_spec(1, width = "4cm")%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  ## stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)",
                  )%>%
    footnote(general ="2018年《中国统计年鉴》重点城市气象数据，中国气象局2017年各城市气象数据库数据整理计算得到",
             general_title = "数据来源：",
             threeparttable = T)

```

根据表 \@ref(tab:tab-wea) 可以发现，2017年35个城市中平均气温最小的是哈尔滨，为
5.13℃，最大的是海口，为24.63℃。2017年35所城市年均降水量973.74毫米，其中最小的是
银川，降水量为211.3毫米，最大的是广州，降水量为2067.4毫米，且由标准差也可以看出，
各城市间2017年降水量差异很大。平均相对湿度最小的是呼和浩特，为44%，最大的是海口，
为82.17%。平均风速和平均气压最小的都是西宁，分别为1.05米/秒、771.49百帕，两者
最大的分别是呼和浩特和上海，取值分别为3.39米/秒、1016.76百帕。全年年均日照时数
2096.13小时，最小的是贵阳，为1100.7小时，最大的是银川，为2894.5小时，同样由标准差
可以看出，各城市间日照时数也存在很大差异。

如前文所述，气温较高、气压较低、降水较多、相对湿度较大、一定范围内风速越大时，空气
质量越好。相对应的，由以上分析可得，35所城市中，海口的平均气温最高，最低气温也最高，
平均相对湿度最大，因此，海口的气象条件有利于污染物的扩散，其空气质量相对最优。同时
可以发现，西宁的平均风速和平均气压在35所城市中都是最小的，结合之前的分析可以知道，
西宁的地区生产总值在35所城市中排名最后，只占各城市年均地区生产总值的13.57%，根据
地区生产总值可以知道其经济发展比较落后，但是其空气质量并没有比某些比它发展的好城市
更好，这也与西宁的气象因素有关，其气象条件不利于污染物的扩散。

#### 地理因素变量描述性分析

地理因素对空气质量的好坏同样具有很大的影响。在我们研究的变量中，地理因素中的主要变量
是是否集中供暖及是否临海。我们知道集中供暖煤炭燃烧的污染物排放是影响空气质量的重要因
素之一，临海城市由于海风的作用也比较利于污染物的扩散。为了探究城市空气质量与地理环境
之间的关系，对2017年35个城市的空气质量相关数据与地理环境相关数据进行了t检验，检
验结果显示在表 \@ref(tab:tab-geo) 中。

```{r tab-geo, eval=T, results='markup'}
varlist <- read.csv("./data/var_list.csv")
dat.y <- read.csv("./data/yearly.csv")

geovar <- names(dat.y)[c(1,3,10,15, 41,42)]
geo.y17 <- dat.y[dat.y$year==2017, geovar]

varinfo <- varlist[match(geovar[-1], as.character(varlist$var_code)), c(1,2,7)]

geo.stat <- as.data.frame(rbind(
apply(geo.y17[2:4], 2, function(x) {temp <- t.test(x~geo.y17$heating)
    c(temp$esti, temp$stat, temp$p.v)
})[c(2,1,3,4),],
apply(geo.y17[2:4], 2, function(x) {temp <- t.test(x~geo.y17$coast)
    c(temp$esti, temp$stat, temp$p.v)
})[c(2,1,3,4),]))

geo.count <- c(as.vector(table(geo.y17[,5]))[2:1],NA,NA,
               as.vector(table(geo.y17[,6]))[2:1],NA,NA)

geo.factor <- c("是否集体供暖", NA,"t统计量值","P值", "是否临海", NA,"t统计量值","P值")
geo.factor2 <- c("是", "否", rep(NA, 2), "是", "否", rep(NA, 2))
geo.stat <- cbind(geo.factor, geo.factor2, geo.count, geo.stat)


names(geo.stat) <- c("属性", "取值", "城市数量", "AQI均值", "为优的平均天数", "严重污染平均天数")
write.csv(geo.stat, "./results/geo.stat.csv", row.names=F)

library("kableExtra")
knitr::kable(geo.stat, row.names =F, align = c("l","l","c", rep("r", 3)),
             digits=2,
             caption="2017年35个城市空气质量与地理环境关系检验",
             longtable = TRUE, booktabs = TRUE, linesep  = c("", "", "", "\\hline"), escape = F) %>%
    kable_styling(full_width = T) %>%
    column_spec(1:3, width = c("3.5cm", "1.1cm", "2cm"))%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  ## stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)",
                  )%>%
    footnote(general ="城市是否集中供暖来自各城市政府网站，城市是否临海根据城市地理位置确定",
             general_title = "数据来源：",
             threeparttable = T)

```

由 \@ref(tab:tab-geo) 可以看出，在我们所研究的35所城市中，集中供暖的有17个，剩下
的18所城市不采取集中供暖。在集中供暖的17所城市中，其年均AQI为99.79，按照空气质量
等级划分标准，其空气质量平均来讲整体为良，但空气质量指数已经几乎接近100，近似为
轻度污染。在集中供暖的城市中，2017年空气质量为优的平均天数为34.82天，严重污染
天数为6.18天。而在没有集中供暖的18所城市中，其空气质量年均指数为72.94，空气质量
为优的平均天数为102.72天，严重污染的平均天数只有0.28天。且根据表 \@ref(tab:tab-geo) 可以发现，在
5%的显著性水平下，检验结果通过显著性检验，说明集中供暖的城市与没有集中供暖的城市
其空气质量存在显著差异。同样，在5%的显著性水平下，临海和不临海的城市的空气质量
之间也存在显著差异。在我们的研究范围内，临海城市有10个，内陆城市有25个，临海城市
的年均AQI为71.21，空气质量为优的平均天数为98.90天，严重污染的平均天数为0.4天。
内陆城市的年均AQI为91.89，空气质量为优的平均天数为58.08天，严重污染的平均天数为
4.24天。以上分析表明，城市空气质量与地理因素有关，是否集中供暖和是否临海其城市的
空气质量差异较大。

## 35城市空气质量指数时间分布特征

### 35城市空气质量年度变化趋势分析

城市空气质量比之前更好还是更差是我们比较关心的问题，也可以对空气污染的治理提供些许
建议。为了进一步观察2014-2017年各城市空气质量的变动趋势，根据2014-2017年各城市的年均
AQI数据绘制了35所城市2014-2017年年均AQI的变化趋势图，如图 \@ref(fig:fig-aqi-year) 所示。

```{r fig-aqi-year, eval=T, fig.height=7.5, fig.width=6.5, out.width="1.05\\textwidth", fig.pos="h", fig.cap = "35个城市2014-2017年平均AQI变化趋势图", dev=c("cairo_pdf"), dev.args=list(family="Microsoft YaHei UI Light")}
varlist <- read.csv("./data/var_list.csv")
dat.y <- read.csv("./data/yearly.csv")

dat.y$city <- factor(as.character(dat.y$city), levels=dat.y$city[(0:34)*4+1])

library(lattice)


xyplot(AQI ~ year | city, data=dat.y,
layout = c(5,7), type = c("o"), 
as.table = TRUE, grid = F,
abline=c(list(h=c(75, 100, 125)),trellis.par.get("reference.line")),
strip = strip.custom(bg="grey", strip.levels = T),
par.strip.text=list(cex=0.8),
digits=0, cex=0.7, cex.axis = 0.8, pch=20,
scales=list(x=list(at=2014:2017, labels=c(14:17)),
            y=list(at=seq(50, 150, 50))),
xlab=list(label="年份",cex=0.9, digits=0), ylab = ""
)

```

图 \@ref(fig:fig-aqi-year) 显示了35所城市2014-2017年年均AQI的变动趋势，每个城市的
趋势图中横轴表示年份，纵轴表示年均AQI。由图 \@ref(fig:fig-aqi-year) 可以清晰地看出，
35所城市的年均AQI变化大致可以分为以下三种类别。

第一种类别是诸如北京、沈阳、长春、南京、济南、青岛、郑州、武汉、长沙、南宁、重庆，
这些城市在2014-2017年年均空气质量指数均呈逐渐下降的趋势，这表明这些城市在这些年份
中空气质量得到了相应的改善。具体来讲，这些城市中，北京、济南这4年AQI均处于100-150
之间，但呈下降趋势。沈阳、武汉在2014-2015年年均AQI处于100-150的范围之内，按空气质量
划分标准，其空气质量为轻度污染，但在从2016年开始，其年均空气质量指数位于50-100之间，
说明其空气质量得到了比较大的改善。同样，长春在2014-2015年其年均AQI接近100，但在后
两年，其年均AQI处于100之下。南京的空气质量指数从2014年的大于100，到2017年小于100。
青岛、长沙、南宁其年均空气质量指数随时间变化均在缓慢下降。郑州的空气质量指数变动
幅度较大，2015年相比2014年有所上升，但之后年份均在下降。2015年之后重庆年均空气质量
指数基本保持不变。整体来讲，这些城市在2014-2017年空气质量逐年得到改善。

第二种类别是诸如天津、石家庄、太原、哈尔滨、合肥、西安、兰州、银川，这些城市2014-2017
年年均AQI均呈上升的趋势，但AQI上升的年份不尽相同。具体来讲，天津、石家庄、太原、合肥、
西安这五所城市在2015年空气质量指数均有所下降，之后年份有不同幅度的上涨。天津、太原、
合肥、西安在2015年空气质量得到相应改善，但之后年份相对2015年空气质量状况变差，其中天津
2015年之后年均空气质量指数有缓慢增长，而太原、西安空气质量指数增长幅度较大，合肥2015年
之后年均空气质量指数变化不大，表明太原、西安2015年之后空气质量在逐年恶化，且其恶化程度
相对较高。石家庄的年均空气质量指数在每年都处于比较高的位置，但其在2015年年均空气质量
指数有较大幅度下降，但之后年份又有小幅增长。哈尔滨在2014-2015年空气质量指数变化不大，
2016年出现较大幅度下降，之后又有一定程度的增加。兰州在2014-2015年、2016-2017年年均空气
质量指数均没有什么大的变化，但2016年相比2015年有一定幅度的上涨。而从银川的年均空气质量
指数变化趋势来看，从2014-2017年，其空气质量指数在逐年上涨，表明其空气质量在逐年恶化。
整体来讲，在2014-2017年，这些城市的空气质量在不同程度的逐年恶化。

第三种类别是诸如呼和浩特、大连、上海、杭州、宁波、福州、厦门、南昌、广州、深圳、海口、
成都、贵阳、昆明、西宁、乌鲁木齐，这些城市2014-2017年年均AQI变动幅度不大，基本处于稳定
状态。具体来讲，呼和浩特、深圳年均AQI在2017年相比之前年份有较小幅度上升，大连、上海、
宁波在2015年相对2014年年均空气质量指数有较小幅度增长，杭州、海口、成都、昆明空气质量
指数基本无显著变化，福州、厦门、南昌、广州在2015年年均空气质量指数稍有下降，2017年又
稍有上涨。贵阳2015年相对2014年年均空气质量指数有所下降，2017年相对2016年其空气质量又
稍微有些好转。西宁、乌鲁木齐2015年相对2014年年均空气质量指数有较小幅度下降。整体来讲，
这些城市年均空气质量指数虽然存在小幅度变化，但极其微小的变动可以认为其空气质量指数基本
保持稳定，2014-2017年这些年份年均空气质量基本稳定在一定的水平。

同时可以根据图 \@ref(fig:fig-aqi-year) 将35所城市的空气质量指数分为三个等级，其中空气
质量较差的城市有北京、天津、石家庄、太原、沈阳、济南、郑州、西安、乌鲁木齐；空气质量
中等的城市有呼和浩特、大连、长春、哈尔滨、上海、南京、杭州、宁波、合肥、南昌、青岛、
武汉、长沙、重庆、成都、兰州、西宁、银川；空气质量较好的城市有福州、厦门、广州、深圳、
南宁、海口、贵阳、昆明。这与之前的分析也保持了一定的一致性。

在对年均空气质量指数有一定的了解之后，我们进一步分析了2014-2017年35个城市空气质量为优
的天数的相关情况，绘制了35所城市2014-2017年空气质量为优的天数的点阵图，如图 \@ref(fig:fig-excel-year) 所示。
在图 \@ref(fig:fig-excel-year) 中，横轴表示空气质量为优的天数，纵轴表示35所城市排序
情况，从上到下依次表示空气质量为优天数最多的城市到空气质量为优天数最少的城市。由于
篇幅所限，图中4、5、6、7分别表示2014年、2015年、2016年、2017年，数字所在位置对应的
横轴即表示对应年份空气质量为优的天数。图中的折线表示2014-2017年各城市空气质量为优
天数的四年平均。

```{r fig-excel-year, eval=T, fig.height=7.5, fig.width=6.5, out.width="1.05\\textwidth", fig.pos="h", fig.cap = "35城市2014-2017年空气质量为优天数点阵图", dev=c("cairo_pdf"), dev.args=list(family="Microsoft YaHei")}
varlist <- read.csv("./data/var_list.csv")
dat.y <- read.csv("./data/yearly.csv")

dat.y$city <- factor(as.character(dat.y$city), levels=dat.y$city[(0:34)*4+1])

library(lattice)

dotplot(reorder(city, excellent) ~ excellent, data=dat.y,
        jitter.y = F, pch = c("4","5", "6", "7"), cex=1,
        xlab = list(label="空气质量为优天数",cex=0.9),
        xlim=c(-15, 280),
        scales=list(x=list(at=seq(0, 270, 30))),
        type = c("p", "a"), lwd=0.5, col=1:4)

```

根据图 \@ref(fig:fig-excel-year) 可以看出，这四年来，空气质量为优的平均天数最多的
是海口，达到了200天左右，而最少的是济南，只有不到10天。同样，直观来看，我们也可以
把空气质量为优的平均天数分为以下几种类别。第一类为空气质量为优的平均天数在200天
左右的城市，很明显只有海口一个城市。第二类为空气质量为优的平均天数在120-180天左右，
这些城市包括深圳、厦门、南宁、昆明、贵阳、福州。第三类为空气质量为优的平均天数在
40-90天左右，这些城市包括南昌、广州、哈尔滨、宁波、大连、长沙、重庆、上海、北京、
呼和浩特。第四类为空气质量为优的平均天数在40天左右的城市，有杭州、长春、南京、合肥、
武汉、青岛。第五类为空气质量为优的天数小于40天的城市，包括乌鲁木齐、沈阳、成都、
西宁、天津、太原、石家庄、西安、银川、兰州、郑州、济南。从以上分析可以看出，在所
研究的35个城市中，大部分城市空气质量为优的平均天数在90天以下，占全年总天数的比例
不到1/4，只有少部分城市空气质量为优的平均天数可以达到全年总天数的1/3。这也显示出
目前中国城市空气质量依然有待提升。

从时间维度看，部分城市空气质量为优的天数在2014年最少，其次是2015年、2016年，在2017年
最多，比如南宁、宁波、长沙、重庆、北京、杭州、长春、南京、合肥、武汉、成都、西宁、西安、
郑州，表明这些城市的空气质量在不断改善。然而，从图 \@ref(fig:fig-excel-year) 中可以发现，也有些城市与这一现象
产生了背离。比如，海口在2014-2016年，空气质量为优的天数依次在增加，但2017年其空气中
质量为优的天数却出现了明显下降，比2016年减少90天左右。深圳在2015年比2014年空气质量为优
的天数相差不大，但2017年比2016年减少25天左右。厦门2015年比2014年空气质量为优的天数增加
80天左右，但2016年却比2015年减少30天左右。同样，昆明、贵阳、天津、石家庄、银川、济南
2016年相对2015年空气质量为优的天数有所下降，福州、南昌、广州、哈尔滨、上海、呼和浩特、
青岛、乌鲁木齐、沈阳、太原、石家庄2017年相对2016年空气质量为优的天数有所下降，宁波、
大连、兰州2015年相对2014年空气质量为优的天数有所下降。整体来看，空气质量为优天数越多
的城市其变动幅度越大，而空气质量为优天数越少的城市其变动幅度相对越小。

值得一提的是，空气质量为优天数的排序与平均AQI的排序并不严格一致，表明一部分地区AQI变化差异
很大，另一部分地区AQI相对较为稳定。以北京、武汉为例，虽然从图 \@ref(fig:fig-excel-year) 中
可以看出，北京空气质量为优的平均天数比武汉多，但是结合之前的分析可以发现，北京的年均AQI却
比武汉的年均AQI大。这就表明，北京的空气质量指数分布比较分散，其空气质量好坏差异较大，而
武汉与其相反，其空气质量指数分布比较集中，其空气质量状况没有特别大的差异，基本处于比较稳定
的水平。其他部分城市与此有类似的现象。

### 35城市空气质量月度变化趋势分析

之前的分析表明了城市空气质量年度变化的整体概况，为了进一步对其进行更加细致的分析，我们
选取了7个代表性城市，对城市空气质量指数的月度变化进行相应分析。在代表性城市的选取中，
首先考虑到的是中国的地理区划。我们知道，中国的地理区划可以分为东北、华东、华北、华南、
西南、西北、华中，各区域有各区域的地理、气候、经济等方面的特点。其次，考虑到各城市的
地理位置，在我们所研究的城市中，有些城市临海，有些城市采取集中供暖的措施，有些城市这
两者都没有。最后，考虑到空气质量的好坏。最终，选取北京、上海、青岛、武汉、深圳、昆明、
西安作为代表性城市进行相应分析。在这七个城市中，北京属于华北地区，上海、青岛属于华东
地区，武汉属于华中地区，深圳属于华南地区，昆明属于西南地区，西安属于西北地区，且北京、
青岛、西安在冬季有集中供暖措施，上海、青岛、深圳是临海城市，而深圳、昆明的空气质量在
所研究的城市中相对较好，北京、上海、青岛、武汉、西安的空气质量在所研究的城市中相对较
差。之后绘制了7个代表城市的月均AQI趋势图，如图 \@ref(fig:fig-aqi-month) 所示。

```{r fig-aqi-month, eval=T, fig.height=7.4,fig.width=6.5,out.width="1.0\\textwidth", fig.pos="h", fig.cap = "7个代表性城市2014-2017年月度平均AQI变化趋势图", dev=c("cairo_pdf"),dev.args=list(family="Microsoft YaHei UI Light")}
varlist <- read.csv("./data/var_list.csv")
dat.m <- read.csv("./data/monthly.csv")

dat.m$city <- factor(as.character(dat.m$city), levels=dat.m$city[(0:34)*48+1])

select.city <- c("北京","上海", "深圳","武汉","西安", "青岛", "昆明")
select.m <- dat.m[dat.m$city %in% select.city,c(1:4,13,18)]
select.m$city <- factor(as.character(select.m$city), levels=select.m$city[(0:34)*48+1])

library(zoo)
aqi.select.m <- tapply(select.m$AQI, select.m$city, rep)
aqi.select.m <- as.zoo(ts(sapply(aqi.select.m, rep), start=c(2014,1), frequency = 12))

library(lattice)
xyplot(aqi.select.m, type="o", pch=20,
       layout = c(1,7), as.table = T,
       ylim=c(20,250),
       strip.left = T, strip = F, grid=T,
       scales=list(x=list(at=2014:2018),
                   y=list(at=c(50, 150, 250))
                   ),
       xlab=list(label="时间",cex=0.9, digits=0),
       ## ylab = "AQI"
       )

```

从7个代表城市2014-2017年的月均AQI趋势图中可以看出，北京、武汉、西安月均AQI趋势变化
最大，昆明月均AQI趋势变化最小。具体来看，北京几乎每年的5、6、7、10月空气质量指数均
会增加，尤其以10月最为明显。上海2016-2017年7、11月空气质量指数会明显增加，武汉每年
11月、次年1月空气质量指数会增加，西安也是11月开始空气质量变差，到次年3月得到缓解。
青岛、深圳、昆明月均AQI变化幅度没有其他城市大。但整体来看，各代表城市月均AQI变动
趋势存在明显的季节效应，直观来看，各代表性城市在冬季及春季的空气质量比较差，而在
夏季和秋季的空气质量状况相对较好。

由上分析可知，各代表性城市月均AQI存在明显的季节效应，因此，为了测定其季节变动，我们
计算了各代表性城市的季节指数。季节指数是一种以相对数表示的季节变动的衡量指标，可以
刻画序列在各月份的典型季节特征。在长期趋势分析里，时间序列的构成要素有长期趋势、季节
变动、循环变动和不规则变动，我们可以用加法模型和乘法模型来进行相应分析。在乘法模型中，
季节指数是以其平均数等于100%为条件而构成的，它反映了某一月份或季度的数值占全年平均
数值的大小。季节变动的程度根据各季节指数与其平均数（100%）的偏差程度来测量。如果序列
没有季节变动，则各期的季节指数应该等于100%；如果某一月份或季度存在明显的季节变化，则
各期的季节指数应该大于或小于100%。季节变动表现为各月的季节指数围绕100%上下波动，表明
各月的情况与全年平均数的相对关系。7个代表城市的季节指数计算结果如表 \@ref(tab:tab-season) 所示。

```{r tab-season, eval=T, results='markup'}
varlist <- read.csv("./data/var_list.csv")
dat.m <- read.csv("./data/monthly.csv")

dat.m$city <- factor(as.character(dat.m$city), levels=dat.m$city[(0:34)*48+1])

select.city <- c("北京","上海", "深圳","武汉","西安", "青岛", "昆明")
select.m <- dat.m[dat.m$city %in% select.city,c(1:4,13,18)]
select.m$city <- factor(as.character(select.m$city), levels=select.m$city[(0:34)*48+1])

aqi.select.m <- tapply(select.m$AQI, select.m$city, rep)
aqi.select.season <- sapply(aqi.select.m,
                            function(x) decompose(ts(x, start=c(2014,1), frequency = 12), type = "multi")$figure)


aqi.select.trend <- sapply(aqi.select.m,
                           function(x) {temp <- decompose(ts(x, start=c(2014,1), frequency = 12), type = "multi")
                               temp$x/temp$seasonal
                           })

library("tseries")

dimnames(aqi.select.season)[[1]] <- paste0(1:12, "月")
write.csv(aqi.select.season, "./results/aqi.select.season.csv", row.names=T)

library("kableExtra")
knitr::kable(aqi.select.season, row.names =T, align = "c",
             digits=2,
             caption="7个代表性城市月度季节指数",
             longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F) %>%
    kable_styling(full_width = T) %>%
    column_spec(1:8, width = "1.4cm")%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  ## stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)",
                  )%>%
    footnote(general ="对每个城市月度平均AQI建立时间序列因素分解乘法模型计算得到",
             general_title = "数据来源：",
             threeparttable = T)

```

表 \@ref(tab:tab-season) 显示的7个代表性城市的月度季节指数以1为平均值，根据表  \@ref(tab:tab-season) 
可以发现，各代表性城市的月度AQI存在些许差异。具体来讲，北京在1、3、5、6、7、10、
11、12月的空气质量指数大于其平均值，分别比平均值高4%、1%、3%、7%、4%、6%、15%、
25%，在2、4、8、9月的空气质量指数小于其平均值，分别比平均值低15%、13%、12%、25%，
表明在北京，12月空气质量状况最差，9月空气质量状况最好。同样，上海在1、4、5、7、
12月空气质量指数大于其平均值，分别比平均值高15%、7%、3%、3%、26%，在2、3、6、8、
9、10、11月的空气质量指数小于其平均值，分别比平均值低1%、8%、3%、14%、12%、15%、
2%，表明上海在10月空气质量最好，在12月空气质量最差，且2、5、6、7、11月空气质量
指数几乎接近平均值。青岛在1、2、4、5、11、12月空气质量指数大于平均值，分别比平均
值高26%、6%、8%、4%、3%、33%，在3、6、7、8、9、10月空气质量指数比其平均值低，分别
比平均值低5%、8%、22%、21%、12%、12%，表明青岛在12月和1月空气质量较差，7月、8月
空气质量较好，且从其差值来看，各月份青岛的空气质量相差较大。武汉在1、2、3、12月
的空气质量指数大于平均值，分别比平均值高45%、23%、5%、31%，在4、6、7、8、9、10、
11月的空气质量指数低于平均值，分别比平均值低15%、23%、25%、14%、11%、6%、10%，其
在5月份的空气质量指数等于平均值，表明武汉在1、2、12月空气质量较差，尤其是1月份，
其空气质量指数比平均值高将近一半，结合原始数据可以发现，这是由于武汉在2014年1月的
空气质量很差，相应的空气质量指数很大，这一个月的数值拉高了所有一月份的平均，同时，
武汉在4、6、7、8月的空气质量相对较好。深圳在1、2、9、10、11、12月的空气质量指数
高于平均值，分别比平均值高13%、8%、10%、16%、5%、15%，在3、4月份和平均值相同，在
5、6、7、8月份低于平均值，分别比平均值低6%、41%、15%、5%，表明深圳在1.、10、12月
空气质量较差，但是在6、7月空气质量较好，尤其以6月为最优，其空气质量指数比6月份
平均值低41%左右。昆明在2、3、4、5、10、11、12月空气质量指数高于平均值，分别比平均
值高3%、16%、12%、15%、3%、1%、11%，在1、6、7、8、9月空气质量指数比平均值低，分别
比平均值低1%、20%、10%、16%、14%，表明昆明在3、4、5、12月空气质量较差，在6、7、8、
9月空气质量较好。西安在1、2、11、12月空气质量指数高于平均值，分别比平均值高56%、
15%、17%、43%，在3-10月空气质量指数低于平均值，分别比平均值低2%、18%、15%、16%、
19%、23%、32%、6%，表明西安在1、2、11、12月空气质量较差，尤其以1、12月最为严重，
在4-9月空气质量较好。这也许与西安的地理位置及火电厂的使用有关。

根据以上分析，可以知道，整体来看，这7个代表性城市也可以分为三大类别。第一类是深圳
和昆明，其空气质量相对较好，相对来讲季节性较弱，且其在6月份空气质量最好。第二类是
北京和上海，存在季节性，且其在12月污染加重。第三类是青岛、武汉和西安，同样存在季节
效应，且其在年底空气污染比较严重，7-9月份空气质量相对其他月份得以改善。综上所述，
我们可以认为，就季节调整之后的空气质量指数来说，城市空气质量的好坏与气象因素及地理
因素具有很大的关系，与社会经济发展的关系不是太大。

从时间维度来看，在7个代表性城市中，除了昆明，其他城市在1月份其空气质量指数均大于
平均值，且西安、武汉最为严重，青岛次之，武汉、西安在1月份的空气质量最差。
相应的，除了北京、上海，其他城市2月份空气质量指数大于其平均值，尤其以武汉最为严重。
在3月份，除上海、青岛之外，其他城市的空气质量指数均大于其平均值，以昆明最为严重。
4月份，除北京、武汉、西安之外，其他城市的空气质量指数均大于其平均值，同样以昆明
最为严重。5月份，除深圳和西安，其他城市的空气质量指数均大于其平均值，同样以昆明
最为严重。6月份，与之前月份有所不同，除北京以外，其他城市的空气质量指数均小于其
平均值，以深圳最为显著。7月份，除了北京和上海，其他城市的空气质量指数均小于其平
均值，以青岛最为显著。8月份所有代表城市的空气质量指数均小于其平均值，表明在8月份
各城市空气质量都达到了最优。9月份，除深圳之外，其他城市空气质量指数均小于其平均值，
以西安最为显著。10月份，除了北京、深圳和昆明，其他城市空气中质量指数均小于其平均值，
以上海最为显著。11月份，除了上海和武汉，其他城市空气质量指数均大于其平均值，以西安、
北京最为显著。12月份所有代表性城市空气质量指数均大于其平均值，表明12月份各城市空气
质量达到最差。综合以上分析，可以发现，不同城市空气质量指数在各月份的分布都不相同，
但有一点一样的是，所有城市在8月份空气质量均达到了最优，在12月份，空气质量均达到了
最差。大部分城市在6月份之后空气质量开始变好，11月份之后，空气质量开始变差。

为了观察城市各月份之间的空气质量指数是否存在相关性，对剔除季节因素之后的月均AQI
进行Ljung-Box检验，检验结果如表 \@ref(tab:tab-box) 所示。结果显示，在5%的显著性水平
下，青岛、武汉、深圳、昆明这四个城市不拒绝原假设，认为这四个城市的月均空气质量指数不
存在自相关，即本月的空气质量不受上月空气质量的影响。而上海、西安这两个城市不拒绝原
假设，认为其月均空气质量指数存在相关，即本月的空气质量与之前月份的空气质量存在一定
的相关性。同时，在10%的显著性水平下，北京的检验结果拒绝原假设，认为其空气质量也与
之前月份的空气质量存在相关性。如果我们将5%的显著性水平定义为强相关，把10%的显著性
水平定义为弱相关，则上海、西安季节调整之后的月均空气质量指数存在强相关性，而
北京季节调增之后的月均空气质量存在弱相关性，其他城市季节调整之后的月均空气质量指数
不存在自相关。

```{r tab-box, eval=T, results='markup'}
varlist <- read.csv("./data/var_list.csv")
dat.m <- read.csv("./data/monthly.csv")

dat.m$city <- factor(as.character(dat.m$city), levels=dat.m$city[(0:34)*48+1])

select.city <- c("北京","上海", "深圳","武汉","西安", "青岛", "昆明")
select.m <- dat.m[dat.m$city %in% select.city,c(1:4,13,18)]
select.m$city <- factor(as.character(select.m$city), levels=select.m$city[(0:34)*48+1])

aqi.select.m <- tapply(select.m$AQI, select.m$city, rep)
aqi.select.season <- sapply(aqi.select.m,
                            function(x) decompose(ts(x, start=c(2014,1), frequency = 12), type = "multi")$figure)

aqi.select.box <- sapply(aqi.select.m,
                           function(x) {temp <- decompose(ts(x, start=c(2014,1), frequency = 12), type = "multi")
                               box.res <- Box.test(temp$x/temp$seasonal, type = "Ljung")
                               c(48, box.res$stat, box.res$p.v)
                           })

temp.name <- dimnames(aqi.select.box)[[2]]
aqi.select.box <- matrix(as.character(round(aqi.select.box,3)),3)
dimnames(aqi.select.box)[[1]] <- c("样本数", "卡方统计量", "P值")
dimnames(aqi.select.box)[[2]] <- temp.name

write.csv(aqi.select.box, "./results/aqi.select.box.csv", row.names=T)

library("kableExtra")
knitr::kable(aqi.select.box, row.names =T, align = "c",
             digits=3,
             caption="7个代表性城市剔除季节性因素后月均AQI自相关检验",
             longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F) %>%
    kable_styling(full_width = T) %>%
    column_spec(1:8, width = c("2.2cm", rep("1.3cm", 7)))%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  ## stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)",
                  )%>%
    footnote(general ="对7个代表性城市剔除季节性因素后月均AQI数据进行 Ljung-Box 检验结果整理得到",
             general_title = "数据来源：",
             threeparttable = T)

```

为了进一步探究上海、西安两个城市季节调整后月均空气质量指数自相关性的阶数，我们对其
差分之后的序列进行自相关性检验。结果显示，上海经季节调整之后的月均空气质量指数存在
一阶自相关，说明经季节调整后上海当月的空气质量指数与上一个月的空气质量指数相关，而
与其他之前月份的空气质量指数无关。而西安经季节调整之后的空气质量指数存在二阶自相关，
说明经季节调整后西安当月的空气质量指数与其前两个月的空气质量指数相关，而与其他之前
月份的空气质量指数无关。

## 35城市空气质量指数空间分布特征

### 城市空气质量空间地理分布
根据以往的研究发现，空气质量存在明显的空间分布特征，为了对所研究城市的空间分布特征
进行更加清晰的描述，在了解了城市空气质量时间分布特征之后，我们对城市空气质量的空间
分布特征进行了分析，根据各城市的经度及纬度，绘制了35个城市2017年年均AQI的空间分布图，
结果如图 \@ref(fig:fig-city-map) 所示。

```{r fig-city-map, eval=T, fig.height=7, fig.width=6.5,out.width="1.0\\textwidth", fig.pos="h", fig.cap = "35个城市2017年平均AQI空间分布图", dev=c("cairo_pdf"), dev.args=list(family="Microsoft YaHei UI Light")}

library(ggplot2)
library(ggrepel)
library(maptools)
mydat <- readShapePoly('./data/mapdata/bou2_4p.shp')
china <- fortify(mydat)

city <- read.csv("./data/city_geo.csv", stringsAsFactors=FALSE)
city$lat <- paste(substr(city$lat,1,2), substr(city$lat,3,4))
city$long <- paste(substr(city$long, 1, nchar(city$long)-2),
                   substr(city$long, nchar(city$long)-1, nchar(city$long)))

city$lat = as.numeric(measurements::conv_unit(city$lat,
                                   from = 'deg_dec_min', to = 'dec_deg'))
city$long = as.numeric(measurements::conv_unit(city$long,
                                               from = 'deg_dec_min', to = 'dec_deg'))

aqi.y17 <- read.csv("./results/aqi.y17.csv")
city$aqi <- aqi.y17[,2]
city$grade <- aqi.y17[,3]
write.csv(city, "./results/city_co.csv")
levels(city$grade) <- c("yellow2", "orange", "green")

ggplot() +
    geom_polygon(data=china, aes(x = long, y = lat, group = group),
                 col = '#FD9FA4', fill = NA, lwd=0.15) +
    geom_point(data = city, aes(x = long, y = lat), cex=city$aqi/20,
               show.legend = F, alpha = 0.8, color = as.character(city$grade)) +
    geom_text_repel(data = city, aes(x = long, y = lat, label = city),
                    cex=3) +
    labs(x = '经度', y = '纬度') + 
    theme_bw()+
    coord_map()+
  theme(text = element_text(),
        ## plot.title = element_text(hjust = 0.5)
        )

```

在图 \@ref(fig:fig-city-map) 中，横轴表示城市经度，纵轴表示城市纬度，图中点的大小
表示2017年年均空气质量指数的大小，点越大，2017年年均空气质量指数越大，点越小，2017
年年均空气质量指数越小。图中点的颜色根据空气质量等级划分标准中的颜色进行定义，绿色
表示空气质量指数为0-50，空气质量等级为优；黄色表示空气质量指数为51-100，空气质量等
级为良；橙色表示空气质量指数为101-150，空气质量等级为轻度污染。

根据图 \@ref(fig:fig-city-map) 可以看出，海口空气质量最好，2017年年均空气质量等级为
优；昆明、贵阳、南宁、深圳、厦门、福州、大连2017年年均空气质量等级为良，但是这些城市
的点的相对较小，说明其年均空气质量指数虽在50-100之间，但是比较偏向50；成都、重庆、
长沙、武汉、南昌、杭州、合肥、南京、上海、宁波、郑州、青岛、西宁、银川、呼和浩特、
沈阳、长春、哈尔滨这些城市在2017年年均空气质量指数也位于50-100之间，但其更接近100；
而乌鲁木齐、兰州、西安、石家庄、北京、天津、太原、济南这些城市的空气质量指数在101-150
之间，表明这些城市2017年空气质量等级平均来讲处于轻度污染。

根据以上分析，我们可以根据地理位置，结合空气质量指数的具体数值，将城市空气质量指数
的空间分布基本归为以下四类。第一类是华北和西北地区的城市，主要包括北京、天津、石家庄、
太原、呼和浩特、西安、兰州、乌鲁木齐、西宁、银川、郑州、青岛、济南。从地理位置来看，
这些城市的明显特征是位于秦岭淮河以北，且从图中可以发现其空气质量相对较差，尤其是乌鲁
木齐、兰州、西安、石家庄、北京、天津、太原、济南，这些城市2017年年均空气质量指数均
超过了100，在所研究的城市中空气质量状况最差。第二类是东北地区的城市，主要包括哈尔滨、长春、
沈阳、大连，这四所城市2017年年均空气质量指数均在50-100之间。从图中可以发现，大连的
空气质量稍比其他三所城市好，也许是由于大连临海的优势地理条件，海风有利于污染物的扩散，
海水也会吸附部分污染物。第三类是长江沿线城市，主要包括成都、重庆、长沙、南昌、武汉、
合肥、杭州、南京、上海、宁波，这些城市的空气质量指数也在50-100之间，但其处于秦岭淮河
以南，且其空气质量相对北方城市稍好一点。第四类是华南地区及部分西南地区城市，主要
包括昆明、贵阳、南宁、广州、深圳、福州、厦门、海口，除广州之外，其余城市空气质量在35
所城市中表现最佳，且海口是2017年唯一一个年均空气质量为优的城市。以上分析表明，空气
质量指数存在明显的空间分布特征，整体来讲，北方城市的空气质量比南方城市的空气质量差，
尤其以华北地区最为严重，且越往南方走，空气质量状况越好。

### 城市空气质量空间相关性分析

空间自相关是检验某一要素的属性值是否显著的与其相邻空间点上的属性值相关联的重要指
标，可以揭示空间参考单元与其临近的空间单元属性特征值之间的相似性或相关性。测度空
间自相关性的指标可以分为全局空间自相关和局域空间自相关，常用的两个指标就是全局Moran's I指数 和局域Moran's I指数。全局空间自相关重点在于整体数值，局部空间自相关
重点在于高高、高低、低低、低高的四种分类状态。

全域空间相关性一般用Moran's I指数进行测度，全域莫兰指数从总体上反映了研究对象的空
间相关性，其计算公式为：

$$\mathrm{I}=\frac{n \sum_{i=1}^{n} \sum_{j=1}^{n} w_{i j}\left(x_{i}-\overline{x}\right)\left(x_{j}-\overline{x}\right)}{\sum_{i=1}^{n} \sum_{j=1}^{n} w_{i j} \sum_{i=1}^{n}\left(x_{i}-\overline{x}\right)^{2}}$$ {#eq:m-4.5.2-1}

其中$i\ne j$， $n$是所研究的空间单元数，${{x}_{i}}$ 和${{x}_{j}}$分别表示某现
象$x$在空间单元$i$和 $j$上的观测值，${{w}_{ij}}$ 是空间权值矩阵，表示$i$区域和
$j$区域之间的邻近关系，可以用距离、二进制连接或$K$最近邻来度量。不同的权值矩阵对
结果会产生不同的影响。

Moran's I指数的取值范围在[-1,1]之间，$\text{I}>0$表示空间正相关，其值越大表明空
间相关性越明显，说明相似的观测值在空间集聚；$\text{I}<0$表示空间负相关，其值越小
表明空间差异性越大，说明相似的观测值在空间分散；$\text{I}=0$表示空间呈随机性，说
明观测值在空间上没有规律。

局域空间相关性可以用局域Moran's I进行测度，对于第 $i$ 个单元，局域莫兰定义为：

$$I_i=\frac{x_i-\bar{x}}{S^2}\sum_{j=1}^nw_{ij}(x_j-\bar{x})$$ {#eq:m-4.5.2-2}

其中， $S^2=1/n \sum_{i=1}^n(x_i-\bar{x})^2 , \bar{x}=1/n\sum_{i=1}^nx_i ,i≠j$ ，其
符号含义与全局莫兰指数相同。正的局域莫兰指数表示一个高值被高值所包围，或一个低值
被低值所包围；负的局域莫兰指数表示一个高值被低值所包围，或一个低值被高值所包围。

在大致了解了35个城市空气质量指数的空间分布特征之后，我们对城市空气质量指数的空间相关
性进行了分析，全局莫兰指数可以看出2014-2017年各城市空气质量指数在整体上存在空间集聚
现象，局域莫兰指数可以看出各城市的空间集聚程度。因此，我们分别计算了2014-2017年35个
城市年均空气质量指数的全局莫兰指数及其局域莫兰指数，计算结果如表 \@ref(tab:tab-global-moran) 
和表 \@ref(tab:tab-local-moran) 所示。

```{r tab-global-moran, eval=T, results='markup'}

varlist <- read.csv("./data/var_list.csv")

dat.y <- read.csv("./data/yearly.csv")
city <- read.csv("./data/city_geo.csv", stringsAsFactors=FALSE)

dat.y$city <- factor(as.character(dat.y$city), levels=dat.y$city[(0:34)*4+1])
dat.y$lat <- rep(city$lat_d, each=4)
dat.y$long <- rep(city$long_d, each=4)

####### 全局 莫兰指数
library(ape)
source("./codes/geodistance.R")
df.cities <- city[,c(1, 5, 6)]
names(df.cities) <- c("name", "lat", "lon")
city.dist.km <- round(GeoDistanceInMetresMatrix(df.cities) / 1000)
city.dists.inv <- 1/city.dist.km
diag(city.dists.inv) <- 0


moran.g <- by(dat.y[,c(3, 30:31)], as.factor(dat.y$year), 
              function(x) Moran.I(x$AQI, city.dists.inv)
)

moran.g <- sapply(moran.g, function(x) c(x$observed, x$p.v))
dimnames(moran.g)[[1]] <- c("Moran's I", "P值")

write.csv(moran.g, "./results/moran.g.csv", row.names=T)

library("kableExtra")
knitr::kable(moran.g, row.names =T, align = "c",
             digits=3,
             caption="2014-2017年 35个城市莫兰指数",
             longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F) %>%
    kable_styling(full_width = T) %>%
    column_spec(1:5, width = c("3cm", rep("2cm", 4)))%>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  ## stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)",
                  )%>%
    footnote(general ="通过全局莫兰指数定义计算得到",
             general_title = "数据来源：",
             threeparttable = T)

```

在计算莫兰指数时，我们的空间权重矩阵根据空间距离进行计算，将各城市的经纬度
转化为球形距离，将距离的倒数作为权重建立空间权重矩阵。根据表 \@ref(tab:tab-global-moran) 
可以看出，2014-2017年35个城市全局空自相关指数分别为：0.176、0.206、0.202、0.195，
且在5%的显著性水平下均通过显著性检验，说明2014-2017年35个城市的空气质量指数确实
存在空间相关。首先我们可以看出2014-2017年，空气质量指数的全局莫兰指数均大于
0，说明存在空间正自相关，表明2014-2017年各城市空气质量指数存在空间集聚现象，空气
质量指数较高的城市临近，空气质量指数较低的城市也相互临近。同时可以发现，2015年
相对2014年来讲，城市空气质量指数的全局莫兰指数增大，表明城市空气质量指数的
空间正自相关性增强，即城市空气质量指数的空间聚集现象越来越明显。从2016年开始，
全局空间自相关指数开始下降，表明城市空气质量指数的空间相关性开始减弱，空间集聚程度
开始降低，但依然存在空间正自相关。另外，各城市空气质量指数的全局莫兰指数相对
较小，表明各城市空气质量指数空间相关性相对较弱，集聚程度较低。

为了更清晰地展现2014-2017年35个城市的年均空气质量指数的空间相关性，我们绘制了
2014-2017年年均AQI的空间相关图，如图 \@ref(fig:fig-cor-log) 所示。图中，横轴表示
城市之间的距离，纵轴表示全局莫兰指数。

```{r fig-cor-log, eval=T, fig.height=6,fig.width=6.5,out.width="1.05\\textwidth", fig.pos="h", fig.cap = "2014-2017 年年度平均AQI空间相关图", dev=c("cairo_pdf"),dev.args=list(family="Microsoft YaHei UI Light")}

varlist <- read.csv("./data/var_list.csv")

dat.y <- read.csv("./data/yearly.csv")
city <- read.csv("./data/city_geo.csv", stringsAsFactors=FALSE)

dat.y$city <- factor(as.character(dat.y$city), levels=dat.y$city[(0:34)*4+1])
dat.y$lat <- rep(city$lat_d, each=4)
dat.y$long <- rep(city$long_d, each=4)

####### correlogram
library(ncf)

par(mfrow = c(2,2))
par(mar = c(4.5, 4.1, 2.8, 2.1))
cor.log <- by(dat.y[,c(2:3, 30:31)], as.factor(dat.y$year), 
              function(x){
                  plot(correlog(x$long, x$lat, z=x$AQI,
                                increment = 100, latlon = T, resamp=500, quiet=TRUE),
                       main = x$year[1], ylab="莫兰指数", xlab="距离(km)", ylim=c(-1.0,1.0))
                  abline(h=0, col="grey")
              })

```

需要说明的是，图中莫兰指数的取值是根据不同距离范围内的城市进行的计算。具体来讲，
我们将距离以100km为间隔，划分为不同的35个区间，0-100km、100-200km、200-300km等直
到最大距离，在每一距离范围内都存在一些城市，根据这一范围内的城市之间的距离计算权
重矩阵，计算这些城市的莫兰指数，相当于计算特定距离范围内的莫兰指数，来反映特定距
离内城市空气质量指数的相关性。由于距离在0-100km及大于3500km之内的城市数量较少，
因此我们剔除掉城市数量较少的几个距离，用剩下的距离内的城市进行莫兰指数的计算。图
中实心黑色点表示莫兰指数在5%的显著性水平下显著，空心点表示莫兰指数在5%的显著性水
平下不显著。根据图 \@ref(fig:fig-cor-log) 可以发现，2014年，距离大概在200-300km、400-500km、
1300-1400km、1800-1900km的城市莫兰指数显著，且距离在200-300km、400-500km的城市其
莫兰指数大于0，表明这一距离范围内城市空气质量指数存在空间正自相关，距离在
1300-1400km、1800-1900km的城市的莫兰指数小于0，表明这一距离范围内城市的空气质量
指数存在空间负相关。同样，2015年距离在200-700km之内每隔100km为一个距离范围内的城
市莫兰指数显著且大于0，表明空气质量指数在这一距离范围内存在空间正自相关，距离在
1700-1800km、1800-1900km、2200-2300km的城市莫兰指数显著且小于0，表明空气质量指数
在这一范围内存在空间负自相关。2016-2017年距离在200-500km、600-700km的城市空气质
量指数存在空间正自相关，距离在1400-1500km、1700-1900km、2000-2100km范围内的城市
存在空间负自相关。根据以上分析可得，距离大概在200-700km范围内的城市空气质量指数
存在空间正自相关，这些距离范围内的城市空气质量指数存在空间集聚现象，距离在
1700-2000km范围内的城市的空气质量指数存在空间负相关，这些距离范围内的城市空气质
量指数存在差异。

之后对城市空气质量指数进行局域自相关处理，在进行局域莫兰指数的计算时，根据以往经验，
距离大于1000km的城市之间几乎不存在空间自相关，所以对于空间权重矩阵的计算，将距离大于
1000km的城市权重设定为0。首先绘制35个城市2017年年均AQI的空间分布图，如图 \@ref(fig:fig-moran-map) 
所示。在图 \@ref(fig:fig-moran-map) 中，红色圆圈代表局域莫兰 I指数大于0的城市，黑色方框代表局域莫兰指数小于0
的城市，红色实心圆和黑色实心方框表示局域莫兰指数显著，红色空心圆和黑色空心方框表示
局域莫兰指数不显著，且圆圈和方框越大表示莫兰指数绝对值越大。由图可以看出，以北京、
天津、石家庄、太原、济南、西安为代表的城市其局域莫兰指数均大于0且显著，表明这些城市2017
年空气质量指数存在空间正自相关，根据原始数据可以发现，这些城市的空气质量指数均较大，
因此这些城市周围城市的空气质量指数也比较大，属于高高聚集区域。相对而言，昆明、贵阳、
南宁、海口、广州、福州、厦门，这些城市的局域莫兰指数为负且显著，表明其2017年空气质量
指数存在空间负相关，这些城市的空气质量指数相对较低，其周边城市的空气质量指数就相对较大。

```{r fig-moran-map, eval=T, fig.height=6.5, fig.width=6.5,out.width="1.05\\textwidth", fig.pos="h", fig.cap = "35个城市2017年平均AQI空间分布图", dev=c("cairo_pdf"), dev.args=list(family="Microsoft YaHei UI Light")}

varlist <- read.csv("./data/var_list.csv")

aqi.y17 <- read.csv("./results/aqi.y17.csv")
city <- read.csv("./data/city_geo.csv", stringsAsFactors=FALSE)
city$aqi <- aqi.y17$年均AQI

library(ncf)

aqi.lisa <- lisa(city$long_d, city$lat_d, city$aqi,
                 latlon = T, neigh=1000, resamp=500, quiet=TRUE)

library(maptools)
mydat <- readShapePoly('./data/mapdata/bou2_4p.shp')
par(mar = c(4.5, 4.1, 2.8, 2.1))
plot(mydat, border="gray20", fill=NA, lwd=0.2, xlim=c(73,135), ylim=c(20,40),
     axes=TRUE, asp = 1.2, xlab="经度", ylab="纬度")
plot(aqi.lisa, negh.mean=FALSE, add = T, lwd=1, cex=0.8)

```

进而对2014-2017年各城市的局域自相关指数进行计算，观察其变化趋势，结果如表 
\@ref(tab:tab-local-moran) 所示。根据表 \@ref(tab:tab-local-moran) 可以发现，
不同年份，部分城市局域莫兰指数为正，部分城市局域莫兰指数为负，且根据显著性
检验结果可知，部分城市存在局域空间自相关，部分城市不存在。具体来讲，北京、
天津、沈阳、杭州、南昌、广州、南宁、海口、重庆、昆明，这些城市的局域莫兰指数
在2014-2017年均为正值。北京的局域空间自相关指数在2015年有所上升，之后有所
下降，各年份均通过显著性检验，说明北京的空气质量指数存在聚集现象，其自身的
空气质量指数较高，空气污染较严重，周边城市的空气质量指数也较高，空气污染也
相对严重。天津与北京类似，同样处于高值聚集区域，其自身及周边城市的空气质量
指数均较高，唯一不同的是，天津在2015年局域空间自相关指数相对2014年有所下降，
之后年份有所上升。沈阳在2014-2015年局域空间自相关指数通过显著性检验，说明沈阳
在这两年空气质量指数也处于高值聚集区域，其自身及周边城市的空气质量指数均较高，
且由其指数值可以发现，2015年相比2014年指数增加较大，但2016-2017年局域空间自
相关指数不显著，说明其在2016-2017年不存在空间聚集现象。昆明与沈阳有相同的情况，
但唯一不同的是昆明的空气质量状况较好，其处于低低聚集区域。杭州、南昌、重庆局域
空间自相关指数未通过显著性检验，说明其不存在空间上的相关性。广州、南宁在2015年
局域空间自相关指数通过显著性检验，说明这两个城市在2015年存在空间集聚现象，且其
值大于0，且结合这两个城市的实际情况，说明其自身及周边城市的空气质量指数均相对
较低，但广州取值相对较小，南宁取值也不太大，说明这种低-低空间集聚现象很弱，且
在其他年份这两个城市均不存在空间集聚现象。海口的局域空间自相关指数相对较高，且
在所有年份均通过显著性检验，说明其空气质量指数存在空间集聚现象，海口城市空气
质量较好，这就表明其自身及周边地区空气质量指数均较低，且这种相关性较强。

再如，石家庄、太原、大连、南京、合肥、济南、青岛、郑州、武汉、成都、西安，这些
城市2014-2017年的局域莫兰指数指数均为负。具体来讲，石家庄、太原、大连、南京、
合肥、青岛、武汉、成都，这些城市局域空间自相关指数均未通过显著性检验，济南、郑州
在2015年局域空间自相关指数通过显著性检验，西安在2017年局域空间自相关指数通过显著
性检验，而这些城市的空气质量指数均较高，局域空间自相关指数为负表明其周边城市的
空气质量指数较低。最后一类城市是在这些年份中既存在正的局域空间自相关指数又存在
负的局域空间自相关指数，其经历了区域集聚的转变。比如，呼和浩特、上海、宁波、福州、
厦门、长沙、兰州、西宁、乌鲁木齐，这些城市的局域空间自相关指数均未通过显著性检验。
长春、哈尔滨、深圳、贵阳在2015年局域空间自相关指数为正，通过显著性检验，表明其
空气质量指数为高高聚集区域，2016年为负，但未通过显著性检验。银川在2016-2017年
局域空间自相关指数为正，且通过显著性检验，表明其在这两年内处于高值聚集区域。

综合以上结果可以发现，在2014-2017年，各城市空气质量指数整体上存在空间自相关，且
具体来看，高高聚集区域的城市包括北京、天津、沈阳、长春（2015）、银川（2016-2017），
低低聚集区域的城市是海口，高低聚集区域的城市包括济南、郑州（2015）、西安（2017）。

```{r tab-local-moran, eval=T, results='markup'}

varlist <- read.csv("./data/var_list.csv")

dat.y <- read.csv("./data/yearly.csv")
city <- read.csv("./data/city_geo.csv", stringsAsFactors=FALSE)

dat.y$city <- factor(as.character(dat.y$city), levels=dat.y$city[(0:34)*4+1])
dat.y$lat <- rep(city$lat_d, each=4)
dat.y$long <- rep(city$long_d, each=4)

library(ncf)

aqi.lisa.y <- by(dat.y[,c(2:3, 30:31)], as.factor(dat.y$year), 
              function(x){
                  temp <- lisa(x$long, x$lat, z=x$AQI,
                               latlon = T, neigh=2000, resamp=500, quiet=TRUE)
                  cbind(temp$correlation, temp$p)
              })


aqi.lisa.y <- cbind(city$city, do.call(cbind.data.frame, aqi.lisa.y))
names(aqi.lisa.y) <- c("城市", "2014指数","P值","2015指数","P值","2016指数","P值", "2017指数","P值")

write.csv(aqi.lisa.y, "./results/aqi.lisa.y.csv", row.names=T)

library("kableExtra")
knitr::kable(aqi.lisa.y, row.names =F, align = "c", caption="2014-2017年 35个城市局部Moran's I指数",
             digits = 3,
             longtable = TRUE, booktabs = TRUE, linesep  = "", escape = F) %>%
    kable_styling(latex_options = c("striped", "scale_down", "repeat_header", "hold_position"),
                  stripe_index = rep(1:5, 4)+rep(c(0,10,20,30), each = 5),
                  repeat_header_text = "(续)")%>%
    footnote(general ="通过局部Moran's I指数定义计算得到",
             general_title = "数据来源：")

```

本章对城市空气质量状况进行了基本描述并分析了其时空分布特征，主要分析了2017年35个
城市的空气质量的基本概况，并从社会经济因素、气象因素、地理因素分析了城市空气质量
影响因素的基本概况，进而对各城市年均空气质量指数及年均空气质量为优的天数进行了年度
变化分析，并选取7个代表性城市分析其月均空气质量指数的变化趋势，进而对其进行季节调整。
之后分析了城市空气质量指数的地理分布特征，计算了全局莫兰指数和局域莫兰指数，并绘制
了相关图形。发现，以秦岭淮河为分界线，整体来说，北方城市的空气质量较差，南方城市的
空气质量相对较好，且地理位置越靠近南方，空气质量越好。以北京、天津、石家庄、太原、
济南、西安为代表的城市空气质量较差，且其存在显著的空间正自相关，以海口、贵阳、昆明、
南宁、福州、厦门为代表的城市空气质量较好，存在显著的空间负相关。但近年来，城市空气
质量还是有随时间逐渐好转的趋势。在了解了基本的空气质量概况之后，为了找到影响城市空
气质量的因素，接下来，本文对城市空气质量的影响因素进行分析。


<!--# 参考文献 {-}-->
[//]: # (\bibliography{Bibfile})
