==> report/08-conclusion.Rmd <==
---
title: "结论"
author: "Jin"
date: "2019-04-15"
css: ./style/markdown.css
autoEqnLabels: true
eqnPrefixTemplate: ($$i$$)
linkReferences: true
bibliography: Bibfile.bib
notice: '@*'
csl: ./style/chinese-gb7714-2005-numeric.csl
link-citations: true
---

```{r setup, echo=F}

################# 第 8 章 R 程序代码  ####################


knitr::opts_knit$set(root.dir = getwd())
knitr::opts_chunk$set(echo = FALSE, results = 'hide')
knitr::opts_chunk$set(warning = FALSE, message=FALSE)
```

```{r prepare}
rm(list=ls())
options(digits=4)
options(scipen=100)
graphics.off()
Sys.setlocale("LC_ALL", "Chinese")
```

# 结论

## 结论

本项目在提出研究背景及选题意义，梳理国内外有关广义可加混合模型及空气质量影响因素
有关文献的基础上，给出了广义可加混合模型的模型背景和模型设定，讨论了广义可加混合
模型的两种参数估计方法：DPQL估计和基于切片抽样MCMC估计。在城市空气质量实证部分，
选取数据范围为重点监测的35个副省级及以上城市，收集了空气质量及其影响因素的
2014-2017年的年度和月度数据，主要包括空气质量指数和经济社会因素，气象因素及地理
因素等24个影响变量。利用数据进行城市空气质量状况描述，并对空气质量指数的时空分布
特征进行分析，通过计算全局空间自相关指数，讨论了空气质量指数的时空格局演变。最后
项目分别从空气质量指数，空气质量为优天数和空气严重污染三个不同角度，建立各种广义
可加混合模型并通过模型比较和选择研究影响空气质量的因素。通过以上研究主要得到以下
几点结论：

1\. 对城市空气质量状况进行了基本描述并分析了其时空分布特征，主要分析了2017年35个
城市的空气质量的基本概况及其与社会经济影响因素，气象因素和地理因素3个方面24个自
变量之间的关系。从时间角度看，整体而言我国大部分城市空气质量还是有随时间逐渐好转
的趋势，但少数城市空气质量有逐年恶化的趋势，如太原，西安，另有部分城市空气质量指
数处于高位稳定状态。空间分析结果发现：以秦岭淮河为分界线，整体来说，北方城市的空
气质量较差，南方城市的空气质量相对较好，且地理位置越靠近南方，空气质量越好。以北
京、天津、石家庄、太原、济南、西安为代表的城市空气质量较差，且其存在显著的空间正
自相关，以海口、贵阳、昆明、南宁、福州、厦门为代表的城市空气质量较好，存在显著的
空间负相关。

2\. 在建模过程中，我们考察的因变量一共有 6 个，分别为年均空气质量指数，月均空气
质量指数，年度空气质量为优天数，月度空气质量为优天数，年度空气严重污染和月度空气
严重污染。对这六个变量进行正态性检验，分别采用 Shapiro-Wilk 检验和 Jarque- Bera
检验，对检验结果分析并考虑数据的特征，最终选取的因变量服从的分布为：年均空气质量
指数（正态分布），年度空气质量为优的天数（取对数后近似正态分布）、年度严重污染的
天数（泊松分布）；月均空气质量指数（取对数后正态分布），月度空气质量为优的天数
（泊松分布）、月度是否发生空气严重污染（二点分布）。

3\. 对于年均空气质量指数，通过建立正态可加混合模型并进行模型比较和选择，最优模型
揭示社会经济影响因素中有废气治理完成投资额，固定资产投资额和汽车拥有量，其中废气
治理完成投资额是正的线性影响，固定资产投资和汽车拥有量与年均空气质量指数之间具有
非线性关系；气候因素中的平均风速和平均气温也对年均空气质量指数有显著性影响，其中
平均风速是线性关系，平均气温具有非线性关系。是否临海，是否供暖和纬度三个地理因素
对年均空气质量指数有线性影响。对于月均空气质量指数，对其取对数后并建立正态可加混
合模型并进行模型比较和选择，最优模型揭示社会经济方面影响因素有废气治理完成投资额，
房屋建筑施工面积和汽车拥有量，且这三个因素均为线性影响；平均气温，降水量，平均风
速和平均气压这四个气象变量都对月均空气质量指数有显著性的线性影响；是否临海，是否
供暖和纬度三个地理因素对月均空气质量指数有影响其中是否临海和是否供暖是线性影响，
纬度与月均空气质量指数直接的关系是非线性的。通过比较年均空气质量指数和月均空气质
量指数之间的影响因素。

4\. 对于年度空气质量为优天数，对其取对数后建立正态可加混合模型并进行模型比较和选
择，最优模型揭示社会经济影响因素有GDP，全社会用电量和废气治理完成投资额，其中全
社会用电量和废气治理完成投资额与年度空气质量为优天数之间具有线性关系，GDP具有非
线性作用；气候因素中的平均相对湿度和平均风速对年度空气质量为优天数有线性影响。地
理因素中只有纬度变量对年度空气质量为优天数有显著性影响，并且关系是非线性的。对于
月度空气质量为优天数，通过建立泊松可加混合模型并进行模型比较和选择，最优模型揭示
社会经济影响因素有房屋建筑施工面积和废气治理完成投资额，并且二者与年度空气质量为
优天数之间均具有线性关系；气候因素中的平均气温，降水量，平均相对湿度和平均风速对
月度空气质量为优天数有线性影响。地理因素中是否供暖和纬度变量对月度空气质量为优天
数有显著性影响，是否供暖是线性关系，纬度是非线性影响。

5\. 对于年度空气严重污染天数，通过建立泊松可加混合模型并进行模型比较和选择，最优
模型揭示社会经济影响因素中的常住人口数是影响年度空气严重污染天数多少的主要因素，
并且这个影响是非线性的，总体而言，年度空气严重污染天数是随着常住人口数的增加而逐
渐增多，但常住人口数超过2000万的特大城市，对于空气严重污染控制的比较好，年度空气
严重污染天数并不多。气候因素中的降水量和平均风速对年度空气严重污染天数有负向线性
影响，而日照时数有非线性影响。地理因素中的是否供暖，是否临海和纬度均对年度空气严
重污染天数有显著性线性影响。对于月度是否出现空气严重污染，通过建立Logistic可加混
合模型并进行模型比较和选择，最优模型揭示社会经济影响因素有常住人口数和私人汽车拥
有量，其中私人汽车拥有量与月度是否出现空气严重污染的概率之间具有负向线性关系，常
住人口数与月度是否出现空气严重污染概率之间具有非线性关系；气候因素中的平均气温，
降水量，平均相对湿度，平均风速和日照时数对月度是否出现空气严重污染的概率局有显著
性影响，其中平均气温是非线性影响，其他4个气候因素是负向线性影响。地理因素中只有
纬度变量对月度是否出现空气严重污染有显著性影响，并且这个影响是非线性的。

## 主要不足和研究展望

1\. 广义可加混合模型理论研究需进一步加强。广义可加混合模型估计方法计算强度较大，
在广义可加模型估计中有可能出现不收敛现象。广义可加混合模型另外一个不足是其结果比
标准线性模型的结果更难于解释。此外，由于本研究关注的一个重心是广义可加混合模型的
应用，而且涉及的模型种类较多，导致对具体模型的讨论可能不够全面和深入。目前，广义
可加混合模型的相关理论仍然不成熟，其估计方法和统计推断有待进一步丰富。此外，在目
前算法下，广义可加混合模型出现不收敛的现象时有发生，如何减少不收敛的发生也有待进
一步研究。最后，由于广义可加混合模型框架包含的模型较多，因此对其框架下一些重要具
体模型可以开展更深入的研究。

2\. 实证部分使用数据的范围较小，仍需进一步扩大，以进一步完善空气质量相关研究。考
虑到数据的可获取性问题，本研究选择的数据为2014-2017年35个重点城市的年度和月度数
据。在收集数据过程中，部分变量的数据有缺失，研究中采取了一定的方法对缺失数据进行
插补。在后续研究中，可以分别从时间和空间两个维度选取更多数据，考虑更长时间段和更
多城市的空气质量变化情况，就可以更完整揭示我国空气质量变化规律及其影响因素。此外，
我们一共考虑了社会经济影响因素，气象因素和地理因素3个方面24个自变量，但部分重要
可能影响因素仍为包含在这24个变量中，需要进一步挖掘可能的影响因素。

3\. 从空气质量内部作用机理进行研究，开展因果性研究。值得注意的是，本研究主要利用统
计模型进行空气质量影响因素研究，所得到的结论是相关性结论，而不是因果性结论，也即
得到的空气质量影响因素只能说明其与空气质量之间有显著性相关关系，而它们之间的具体
作用机制，内在因果关系并不清楚。这方面的研究需要结合环境科学，气象学和地理学等学
科进行交叉学科研究。


\cleardoublepage

\pagestyle{emptypage}

\renewcommand{\chapterlabel}{\bibname}

<!-- # 参考文献 {-} -->
[//]: # (\bibliography{Bibfile})
